@@ -1,20 +1,17 @@
 using ApiServer.Model;
 using Microsoft.EntityFrameworkCore;
 
-namespace ApiServer.Data
+namespace ApiServer.Data;
+
+public class DataEventRecordContext : DbContext
 {
-    public class DataEventRecordContext : DbContext
-    {
-        public DataEventRecordContext(DbContextOptions<DataEventRecordContext> options) : base(options)
-        {
-        }
-        
-        public DbSet<DataEventRecord> DataEventRecords { get; set; }
-      
-        protected override void OnModelCreating(ModelBuilder builder)
-        { 
-            builder.Entity<DataEventRecord>().HasKey(m => m.Id); 
-            base.OnModelCreating(builder); 
-        } 
-    }
+    public DataEventRecordContext(DbContextOptions<DataEventRecordContext> options) : base(options) { }
+
+    public DbSet<DataEventRecord> DataEventRecords => Set<DataEventRecord>();
+
+    protected override void OnModelCreating(ModelBuilder builder)
+    { 
+        builder.Entity<DataEventRecord>().HasKey(m => m.Id); 
+        base.OnModelCreating(builder); 
+    } 
 }
