@@ -1,8 +1,7 @@
-﻿namespace ApiServer.Providers
+﻿namespace ApiServer.Providers;
+
+public class NewsGroup
 {
-    public class NewsGroup
-    {
-        public long Id { get; set; }
-        public string Name { get; set; }
-    }
+    public int Id { get; set; }
+    public string Name { get; set; } = string.Empty;
 }
