@@ -1,8 +1,7 @@
-﻿namespace ApiServer.SignalRHubs
+﻿namespace ApiServer.SignalRHubs;
+
+public class UserInfo
 {
-    public class UserInfo
-    {
-        public string ConnectionId { get; set; }
-        public string UserName { get; set; }
-    }
+    public string ConnectionId { get; set; } = string.Empty;
+    public string UserName { get; set; } = string.Empty;
 }
