@@ -2,49 +2,46 @@
 using Microsoft.AspNetCore.Identity;
 using Microsoft.Extensions.Options;
 using StsServerIdentity.Models;
-using System.Collections.Generic;
 using System.Security.Claims;
-using System.Threading.Tasks;
 
-namespace StsServerIdentity
+namespace StsServerIdentity;
+
+public class AdditionalUserClaimsPrincipalFactory : UserClaimsPrincipalFactory<ApplicationUser, IdentityRole>
 {
-    public class AdditionalUserClaimsPrincipalFactory : UserClaimsPrincipalFactory<ApplicationUser, IdentityRole>
+    public AdditionalUserClaimsPrincipalFactory(
+        UserManager<ApplicationUser> userManager,
+        RoleManager<IdentityRole> roleManager,
+        IOptions<IdentityOptions> optionsAccessor)
+        : base(userManager, roleManager, optionsAccessor)
     {
-        public AdditionalUserClaimsPrincipalFactory(
-            UserManager<ApplicationUser> userManager,
-            RoleManager<IdentityRole> roleManager,
-            IOptions<IdentityOptions> optionsAccessor)
-            : base(userManager, roleManager, optionsAccessor)
+    }
+
+    public async override Task<ClaimsPrincipal> CreateAsync(ApplicationUser user)
+    {
+        var principal = await base.CreateAsync(user);
+        var identity = (ClaimsIdentity)principal.Identity;
+
+        var claims = new List<Claim>
         {
+            new Claim(JwtClaimTypes.Role, "dataEventRecords"),
+            new Claim(JwtClaimTypes.Role, "dataEventRecords.user")
+        };
+
+        if (user.DataEventRecordsRole == "dataEventRecords.admin")
+        {
+            claims.Add(new Claim(JwtClaimTypes.Role, "dataEventRecords.admin"));
         }
 
-        public async override Task<ClaimsPrincipal> CreateAsync(ApplicationUser user)
+        if (user.IsAdmin)
         {
-            var principal = await base.CreateAsync(user);
-            var identity = (ClaimsIdentity)principal.Identity;
-
-            var claims = new List<Claim>
-            {
-                new Claim(JwtClaimTypes.Role, "dataEventRecords"),
-                new Claim(JwtClaimTypes.Role, "dataEventRecords.user")
-            };
-
-            if (user.DataEventRecordsRole == "dataEventRecords.admin")
-            {
-                claims.Add(new Claim(JwtClaimTypes.Role, "dataEventRecords.admin"));
-            }
-
-            if (user.IsAdmin)
-            {
-                claims.Add(new Claim(JwtClaimTypes.Role, "admin"));
-            }
-            else
-            {
-                claims.Add(new Claim(JwtClaimTypes.Role, "user"));
-            }
-
-            identity.AddClaims(claims);
-            return principal;
+            claims.Add(new Claim(JwtClaimTypes.Role, "admin"));
         }
+        else
+        {
+            claims.Add(new Claim(JwtClaimTypes.Role, "user"));
+        }
+
+        identity.AddClaims(claims);
+        return principal;
     }
 }
