@@ -4,131 +4,130 @@
 using IdentityServer4.Models;
 using System.Collections.Generic;
 
-namespace StsServerIdentity
+namespace StsServerIdentity;
+
+public class Config
 {
-    public class Config
+    public static IEnumerable<IdentityResource> GetIdentityResources()
     {
-        public static IEnumerable<IdentityResource> GetIdentityResources()
+        return new List<IdentityResource>
         {
-            return new List<IdentityResource>
-            {
-                new IdentityResources.OpenId(),
-                new IdentityResources.Profile(),
-                new IdentityResources.Email(),
-                new IdentityResource("dataeventrecordsir",new []{ "role", "admin", "user", "dataEventRecords", "dataEventRecords.admin" , "dataEventRecords.user" } ),
-            };
-        }
+            new IdentityResources.OpenId(),
+            new IdentityResources.Profile(),
+            new IdentityResources.Email(),
+            new IdentityResource("dataeventrecordsir",new []{ "role", "admin", "user", "dataEventRecords", "dataEventRecords.admin" , "dataEventRecords.user" } ),
+        };
+    }
 
-        public static IEnumerable<ApiScope> GetApiScopes()
+    public static IEnumerable<ApiScope> GetApiScopes()
+    {
+        return new List<ApiScope>
         {
-            return new List<ApiScope>
-            {
-                new ApiScope("dataEventRecords", "Scope for the dataEventRecords ApiResource",
-                    new List<string> { "role", "admin", "user", "dataEventRecords", "dataEventRecords.admin", "dataEventRecords.user"})
-            };
-        }
+            new ApiScope("dataEventRecords", "Scope for the dataEventRecords ApiResource",
+                new List<string> { "role", "admin", "user", "dataEventRecords", "dataEventRecords.admin", "dataEventRecords.user"})
+        };
+    }
 
-        public static IEnumerable<ApiResource> GetApiResources()
+    public static IEnumerable<ApiResource> GetApiResources()
+    {
+        return new List<ApiResource>
         {
-            return new List<ApiResource>
+            new ApiResource("dataEventRecordsApi")
             {
-                new ApiResource("dataEventRecordsApi")
+                ApiSecrets =
                 {
-                    ApiSecrets =
-                    {
-                        new Secret("dataEventRecordsSecret".Sha256())
-                    },
-                    Scopes = new List<string> { "dataEventRecords" }
-                }
-            };
-        }
-
-        // clients want to access resources (aka scopes)
-        public static IEnumerable<Client> GetClients()
-        {
-            // client credentials client
-            return new List<Client>
-            {
-                new Client
-                {
-                    ClientName = "angularclient",
-                    ClientId = "angularclient",
-                    AccessTokenType = AccessTokenType.Jwt,
-                    AccessTokenLifetime = 3300,// 330 seconds, default 60 minutes
-                    IdentityTokenLifetime = 3000,
-
-                    RequireClientSecret = false,
-                    AllowedGrantTypes = GrantTypes.Code,
-                    RequirePkce = true,
-
-                    AllowAccessTokensViaBrowser = true,
-                    AllowOfflineAccess =true,
-                    AlwaysIncludeUserClaimsInIdToken = true,
-                    RedirectUris = new List<string>
-                    {
-                        "https://localhost:44311",
-                        "https://localhost:44311/silent-renew.html"
-
-                    },
-                    PostLogoutRedirectUris = new List<string>
-                    {
-                        "https://localhost:44311/unauthorized",
-                        "https://localhost:44311"
-                    },
-                    AllowedCorsOrigins = new List<string>
-                    {
-                        "https://localhost:44311"
-                    },
-                    AllowedScopes = new List<string>
-                    {
-                        "openid",
-                        "dataEventRecords",
-                        "dataeventrecordsscope",
-                        "role",
-                        "profile",
-                        "email"
-                    }
+                    new Secret("dataEventRecordsSecret".Sha256())
                 },
-                new Client
+                Scopes = new List<string> { "dataEventRecords" }
+            }
+        };
+    }
+
+    // clients want to access resources (aka scopes)
+    public static IEnumerable<Client> GetClients()
+    {
+        // client credentials client
+        return new List<Client>
+        {
+            new Client
+            {
+                ClientName = "angularclient",
+                ClientId = "angularclient",
+                AccessTokenType = AccessTokenType.Jwt,
+                AccessTokenLifetime = 3300,// 330 seconds, default 60 minutes
+                IdentityTokenLifetime = 3000,
+
+                RequireClientSecret = false,
+                AllowedGrantTypes = GrantTypes.Code,
+                RequirePkce = true,
+
+                AllowAccessTokensViaBrowser = true,
+                AllowOfflineAccess =true,
+                AlwaysIncludeUserClaimsInIdToken = true,
+                RedirectUris = new List<string>
                 {
-                    ClientName = "angularclient2",
-                    ClientId = "angularclient2",
-                    AccessTokenType = AccessTokenType.Jwt,
-                    AccessTokenLifetime = 3300,// 330 seconds, default 60 minutes
-                    IdentityTokenLifetime = 3000,
+                    "https://localhost:44311",
+                    "https://localhost:44311/silent-renew.html"
 
-                    RequireClientSecret = false,
-                    AllowedGrantTypes = GrantTypes.Code,
-                    RequirePkce = true,
-
-                    AllowAccessTokensViaBrowser = true,
-                    AllowOfflineAccess =true,
-                    AlwaysIncludeUserClaimsInIdToken = true,
-                    RedirectUris = new List<string>
-                    {
-                        "https://localhost:44395",
-                        "https://localhost:44395/silent-renew.html"
-                    },
-                    PostLogoutRedirectUris = new List<string>
-                    {
-                        "https://localhost:44395/unauthorized",
-                        "https://localhost:44395"
-                    },
-                    AllowedCorsOrigins = new List<string>
-                    {
-                        "https://localhost:44395"
-                    },
-                    AllowedScopes = new List<string>
-                    {
-                        "openid",
-                        "dataEventRecords",
-                        "dataeventrecordsscope",
-                        "role",
-                        "profile",
-                        "email"
-                    }
+                },
+                PostLogoutRedirectUris = new List<string>
+                {
+                    "https://localhost:44311/unauthorized",
+                    "https://localhost:44311"
+                },
+                AllowedCorsOrigins = new List<string>
+                {
+                    "https://localhost:44311"
+                },
+                AllowedScopes = new List<string>
+                {
+                    "openid",
+                    "dataEventRecords",
+                    "dataeventrecordsscope",
+                    "role",
+                    "profile",
+                    "email"
                 }
-            };
-        }
+            },
+            new Client
+            {
+                ClientName = "angularclient2",
+                ClientId = "angularclient2",
+                AccessTokenType = AccessTokenType.Jwt,
+                AccessTokenLifetime = 3300,// 330 seconds, default 60 minutes
+                IdentityTokenLifetime = 3000,
+
+                RequireClientSecret = false,
+                AllowedGrantTypes = GrantTypes.Code,
+                RequirePkce = true,
+
+                AllowAccessTokensViaBrowser = true,
+                AllowOfflineAccess =true,
+                AlwaysIncludeUserClaimsInIdToken = true,
+                RedirectUris = new List<string>
+                {
+                    "https://localhost:44395",
+                    "https://localhost:44395/silent-renew.html"
+                },
+                PostLogoutRedirectUris = new List<string>
+                {
+                    "https://localhost:44395/unauthorized",
+                    "https://localhost:44395"
+                },
+                AllowedCorsOrigins = new List<string>
+                {
+                    "https://localhost:44395"
+                },
+                AllowedScopes = new List<string>
+                {
+                    "openid",
+                    "dataEventRecords",
+                    "dataeventrecordsscope",
+                    "role",
+                    "profile",
+                    "email"
+                }
+            }
+        };
     }
 }
