@@ -1,12 +1,8 @@
-﻿using System;
-using System.Linq;
-using System.Security.Claims;
-using System.Threading.Tasks;
+﻿using System.Security.Claims;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
-using Microsoft.Extensions.Logging;
 using StsServerIdentity.Models.AccountViewModels;
 using StsServerIdentity.Models;
 using IdentityServer4.Services;
@@ -23,763 +19,762 @@ using Microsoft.AspNetCore.WebUtilities;
 using System.Text;
 using System.Text.Encodings.Web;
 
-namespace StsServerIdentity.Controllers
+namespace StsServerIdentity.Controllers;
+
+[Authorize]
+public class AccountController : Controller
 {
-    [Authorize]
-    public class AccountController : Controller
+    private readonly Fido2Store _fido2Storage;
+    private readonly UserManager<ApplicationUser> _userManager;
+    private readonly SignInManager<ApplicationUser> _signInManager;
+    private readonly IEmailSender _emailSender;
+    private readonly ILogger _logger;
+    private readonly IIdentityServerInteractionService _interaction;
+    private readonly IClientStore _clientStore;
+    private readonly IPersistedGrantService _persistedGrantService;
+    private readonly IStringLocalizer _sharedLocalizer;
+    private readonly IAuthenticationSchemeProvider _schemeProvider;
+    private readonly IEventService _events;
+
+    public AccountController(
+        UserManager<ApplicationUser> userManager,
+        IPersistedGrantService persistedGrantService,
+        SignInManager<ApplicationUser> signInManager,
+        IEmailSender emailSender,
+        ILoggerFactory loggerFactory,
+        IIdentityServerInteractionService interaction,
+        IClientStore clientStore,
+        IStringLocalizerFactory factory,
+        Fido2Store fido2Storage,
+        IAuthenticationSchemeProvider schemeProvider,
+        IEventService events)
+
     {
-        private readonly Fido2Storage _fido2Storage;
-        private readonly UserManager<ApplicationUser> _userManager;
-        private readonly SignInManager<ApplicationUser> _signInManager;
-        private readonly IEmailSender _emailSender;
-        private readonly ILogger _logger;
-        private readonly IIdentityServerInteractionService _interaction;
-        private readonly IClientStore _clientStore;
-        private readonly IPersistedGrantService _persistedGrantService;
-        private readonly IStringLocalizer _sharedLocalizer;
-        private readonly IAuthenticationSchemeProvider _schemeProvider;
-        private readonly IEventService _events;
+        _fido2Storage = fido2Storage;
+        _userManager = userManager;
+        _persistedGrantService = persistedGrantService;
+        _signInManager = signInManager;
+        _emailSender = emailSender;
+        _logger = loggerFactory.CreateLogger<AccountController>();
+        _interaction = interaction;
+        _clientStore = clientStore;
+        _schemeProvider = schemeProvider;
+        _events = events;
 
-        public AccountController(
-            UserManager<ApplicationUser> userManager,
-            IPersistedGrantService persistedGrantService,
-            SignInManager<ApplicationUser> signInManager,
-            IEmailSender emailSender,
-            ILoggerFactory loggerFactory,
-            IIdentityServerInteractionService interaction,
-            IClientStore clientStore,
-            IStringLocalizerFactory factory,
-            Fido2Storage fido2Storage,
-            IAuthenticationSchemeProvider schemeProvider,
-            IEventService events)
+        var type = typeof(SharedResource);
+        var assemblyName = new AssemblyName(type.GetTypeInfo().Assembly.FullName);
+        _sharedLocalizer = factory.Create("SharedResource", assemblyName.Name);
+    }
 
+    [HttpGet]
+    [AllowAnonymous]
+    public async Task<IActionResult> Login(string returnUrl)
+    {
+        // build a model so we know what to show on the login page
+        var vm = await BuildLoginViewModelAsync(returnUrl);
+
+        if (vm.IsExternalLoginOnly)
         {
-            _fido2Storage = fido2Storage;
-            _userManager = userManager;
-            _persistedGrantService = persistedGrantService;
-            _signInManager = signInManager;
-            _emailSender = emailSender;
-            _logger = loggerFactory.CreateLogger<AccountController>();
-            _interaction = interaction;
-            _clientStore = clientStore;
-            _schemeProvider = schemeProvider;
-            _events = events;
-
-            var type = typeof(SharedResource);
-            var assemblyName = new AssemblyName(type.GetTypeInfo().Assembly.FullName);
-            _sharedLocalizer = factory.Create("SharedResource", assemblyName.Name);
+            // we only have one option for logging in and it's an external provider
+            return ExternalLogin(vm.ExternalProviders.First().AuthenticationScheme, returnUrl);
         }
 
-        [HttpGet]
-        [AllowAnonymous]
-        public async Task<IActionResult> Login(string returnUrl)
+        return View(vm);
+    }
+
+    //
+    // POST: /Account/Login
+    [HttpPost]
+    [AllowAnonymous]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> Login(LoginInputModel model)
+    {
+        var returnUrl = model.ReturnUrl;
+        var context = await _interaction.GetAuthorizationContextAsync(returnUrl);
+        var requires2Fa = context?.AcrValues.Count(t => t.Contains("mfa")) >= 1;
+
+        var user = await _userManager.FindByNameAsync(model.Email);
+        if (user != null && !user.TwoFactorEnabled && requires2Fa)
         {
-            // build a model so we know what to show on the login page
-            var vm = await BuildLoginViewModelAsync(returnUrl);
-
-            if (vm.IsExternalLoginOnly)
-            {
-                // we only have one option for logging in and it's an external provider
-                return ExternalLogin(vm.ExternalProviders.First().AuthenticationScheme, returnUrl);
-            }
-
-            return View(vm);
+            return RedirectToAction(nameof(ErrorEnable2FA));
         }
 
-        //
-        // POST: /Account/Login
-        [HttpPost]
-        [AllowAnonymous]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Login(LoginInputModel model)
+        ViewData["ReturnUrl"] = returnUrl;
+        if (ModelState.IsValid)
         {
-            var returnUrl = model.ReturnUrl;
-            var context = await _interaction.GetAuthorizationContextAsync(returnUrl);
-            var requires2Fa = context?.AcrValues.Count(t => t.Contains("mfa")) >= 1;
-
-            var user = await _userManager.FindByNameAsync(model.Email);
-            if (user != null && !user.TwoFactorEnabled && requires2Fa)
-            {
-                return RedirectToAction(nameof(ErrorEnable2FA));
-            }
-
-            ViewData["ReturnUrl"] = returnUrl;
-            if (ModelState.IsValid)
-            {
-                // This doesn't count login failures towards account lockout
-                // To enable password failures to trigger account lockout, set lockoutOnFailure: true
-                var result = await _signInManager.PasswordSignInAsync(model.Email, model.Password, model.RememberLogin, lockoutOnFailure: false);
-                if (result.Succeeded)
-                {
-                    _logger.LogInformation(1, "User logged in.");
-                    return RedirectToLocal(returnUrl);
-                }
-                if (result.RequiresTwoFactor)
-                {
-                    var fido2ItemExistsForUser = await _fido2Storage.GetCredentialsByUsername(model.Email);
-                    if (fido2ItemExistsForUser.Count > 0)
-                    {
-                        return RedirectToAction(nameof(LoginFido2Mfa), new { ReturnUrl = returnUrl, RememberMe = model.RememberLogin });
-                    }
-                    else
-                    {
-                        return RedirectToAction(nameof(VerifyCode), new { ReturnUrl = returnUrl, RememberMe = model.RememberLogin });
-                    }
-                }
-
-                if (result.IsLockedOut)
-                {
-                    _logger.LogWarning(2, "User account locked out.");
-                    return View("Lockout");
-                }
-                else
-                {
-                    ModelState.AddModelError(string.Empty, _sharedLocalizer["INVALID_LOGIN_ATTEMPT"]);
-                    return View(await BuildLoginViewModelAsync(model));
-                }
-            }
-
-            // If we got this far, something failed, redisplay form
-            return View(await BuildLoginViewModelAsync(model));
-        }
-
-        [HttpGet]
-        [AllowAnonymous]
-        public async Task<IActionResult> LoginFido2Mfa(string provider, bool rememberMe, string returnUrl = null)
-        {
-            // Require that the user has already logged in via username/password or external login
-            var user = await _signInManager.GetTwoFactorAuthenticationUserAsync();
-            if (user == null)
-            {
-                return View("Error");
-            }
-
-            if (string.IsNullOrEmpty(provider))
-            {
-                provider = "fido2";
-            }
-
-            return View(new MfaModel { /*Provider = provider,*/ ReturnUrl = returnUrl, RememberMe = rememberMe });
-        }
-
-        [HttpGet]
-        [AllowAnonymous]
-        public IActionResult ErrorEnable2FA()
-        {
-            return View();
-        }
-
-        [HttpGet]
-        public async Task<IActionResult> Logout(string logoutId)
-        {
-            // build a model so the logout page knows what to display
-            var vm = await BuildLogoutViewModelAsync(logoutId);
-
-            if (vm.ShowLogoutPrompt == false)
-            {
-                // if the request for logout was properly authenticated from IdentityServer, then
-                // we don't need to show the prompt and can just log the user out directly.
-                return await Logout(vm);
-            }
-
-            return View(vm);
-        }
-
-        /// <summary>
-        /// Handle logout page postback
-        /// </summary>
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Logout(LogoutViewModel model)
-        {
-            var idp = User?.FindFirst(JwtClaimTypes.IdentityProvider)?.Value;
-            var subjectId = HttpContext.User.Identity.GetSubjectId();
-
-            if (idp != null && idp != IdentityServerConstants.LocalIdentityProvider)
-            {
-                if (model.LogoutId == null)
-                {
-                    // if there's no current logout context, we need to create one
-                    // this captures necessary info from the current logged in user
-                    // before we signout and redirect away to the external IdP for signout
-                    model.LogoutId = await _interaction.CreateLogoutContextAsync();
-                }
-
-                // string url = "/Account/Logout?logoutId=" + model.LogoutId;
-                try
-                {
-                    await _signInManager.SignOutAsync();
-                    // await HttpContext.Authentication.SignOutAsync(idp, new AuthenticationProperties { RedirectUri = url });
-                }
-                catch (NotSupportedException)
-                {
-                }
-            }
-
-            // delete authentication cookie
-            await _signInManager.SignOutAsync();
-
-            // set this so UI rendering sees an anonymous user
-            HttpContext.User = new ClaimsPrincipal(new ClaimsIdentity());
-
-            // get context information (client name, post logout redirect URI and iframe for federated signout)
-            var logout = await _interaction.GetLogoutContextAsync(model.LogoutId);
-
-            var vm = new LoggedOutViewModel
-            {
-                AutomaticRedirectAfterSignOut = AccountOptions.AutomaticRedirectAfterSignOut,
-                PostLogoutRedirectUri = logout?.PostLogoutRedirectUri,
-                ClientName = string.IsNullOrEmpty(logout?.ClientName) ? logout?.ClientId : logout?.ClientName,
-                SignOutIframeUrl = logout?.SignOutIFrameUrl,
-                LogoutId = model.LogoutId
-            };
-
-
-            await _persistedGrantService.RemoveAllGrantsAsync(subjectId, "angular2client");
-
-            return View("LoggedOut", vm);
-        }
-
-        //
-        // GET: /Account/Register
-        [HttpGet]
-        [AllowAnonymous]
-        public IActionResult Register(string returnUrl = null)
-        {
-            ViewData["ReturnUrl"] = returnUrl;
-            return View();
-        }
-
-        //
-        // POST: /Account/Register
-        [HttpPost]
-        [AllowAnonymous]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Register(RegisterViewModel model, string returnUrl = null)
-        {
-            ViewData["ReturnUrl"] = returnUrl;
-            if (ModelState.IsValid)
-            {
-                var user = new ApplicationUser
-                {
-                    UserName = model.Email,
-                    Email = model.Email,
-                    IsAdmin = false
-                };
-                var result = await _userManager.CreateAsync(user, model.Password);
-                if (result.Succeeded)
-                {
-                    //var code = await _userManager.GenerateEmailConfirmationTokenAsync(user);
-                    // WebEncoders.Base64UrlEncode(Encoding.UTF8.GetBytes(code));
-                    //var callbackUrl = Url.Action("ConfirmEmail", "Account", new { userId = user.Id, code = code }, protocol: HttpContext.Request.Scheme);
-                    //await _emailSender.SendEmailAsync(model.Email, "Confirm your account",
-                    //    $"Please confirm your account by clicking this link: <a href='{HtmlEncoder.Default.Encode(callbackUrl)}'>link</a>");
-                    //await _signInManager.SignInAsync(user, isPersistent: false);
-                    //_logger.LogInformation(3, "User created a new account with password.");
-                    return RedirectToLocal(returnUrl);
-                }
-                AddErrors(result);
-            }
-
-            // If we got this far, something failed, redisplay form
-            return View(model);
-        }
-
-        //
-        // POST: /Account/ExternalLogin
-        //[ValidateAntiForgeryToken]
-        [HttpPost]
-        [HttpGet]
-        [AllowAnonymous]
-        public IActionResult ExternalLogin(string provider, string returnUrl = null)
-        {
-            // Request a redirect to the external login provider.
-            var redirectUrl = Url.Action("ExternalLoginCallback", "Account", new { ReturnUrl = returnUrl });
-            var properties = _signInManager.ConfigureExternalAuthenticationProperties(provider, redirectUrl);
-            return Challenge(properties, provider);
-        }
-
-        //
-        // GET: /Account/ExternalLoginCallback
-        [HttpGet]
-        [AllowAnonymous]
-        public async Task<IActionResult> ExternalLoginCallback(string returnUrl = null, string remoteError = null)
-        {
-            var context = await _interaction.GetAuthorizationContextAsync(returnUrl);
-            var requires2Fa = context?.AcrValues.Count(t => t.Contains("mfa")) >= 1;
-
-            if (remoteError != null)
-            {
-                ModelState.AddModelError(string.Empty, _sharedLocalizer["EXTERNAL_PROVIDER_ERROR", remoteError]);
-                return View(nameof(Login));
-            }
-            var info = await _signInManager.GetExternalLoginInfoAsync();
-
-            if (info == null)
-            {
-                return RedirectToAction(nameof(Login));
-            }
-
-            var email = info.Principal.FindFirstValue(ClaimTypes.Email);
-
-            if (!string.IsNullOrEmpty(email))
-            {
-                var user = await _userManager.FindByNameAsync(email);
-                if (user != null && !user.TwoFactorEnabled && requires2Fa)
-                {
-                    return RedirectToAction(nameof(ErrorEnable2FA));
-                }
-            }
-
-            // Sign in the user with this external login provider if the user already has a login.
-            var result = await _signInManager.ExternalLoginSignInAsync(info.LoginProvider, info.ProviderKey, isPersistent: false);
+            // This doesn't count login failures towards account lockout
+            // To enable password failures to trigger account lockout, set lockoutOnFailure: true
+            var result = await _signInManager.PasswordSignInAsync(model.Email, model.Password, model.RememberLogin, lockoutOnFailure: false);
             if (result.Succeeded)
             {
-                _logger.LogInformation(5, "User logged in with {Name} provider.", info.LoginProvider);
+                _logger.LogInformation(1, "User logged in.");
                 return RedirectToLocal(returnUrl);
             }
             if (result.RequiresTwoFactor)
             {
-                var fido2ItemExistsForUser = await _fido2Storage.GetCredentialsByUsername(email);
+                var fido2ItemExistsForUser = await _fido2Storage.GetCredentialsByUserNameAsync(model.Email);
                 if (fido2ItemExistsForUser.Count > 0)
                 {
-                    return RedirectToAction(nameof(LoginFido2Mfa), new { ReturnUrl = returnUrl });
+                    return RedirectToAction(nameof(LoginFido2Mfa), new { ReturnUrl = returnUrl, RememberMe = model.RememberLogin });
                 }
                 else
                 {
-                    return RedirectToAction(nameof(SendCode), new { ReturnUrl = returnUrl });
+                    return RedirectToAction(nameof(VerifyCode), new { ReturnUrl = returnUrl, RememberMe = model.RememberLogin });
                 }
             }
+
             if (result.IsLockedOut)
             {
+                _logger.LogWarning(2, "User account locked out.");
                 return View("Lockout");
             }
             else
             {
-                // If the user does not have an account, then ask the user to create an account.
-                ViewData["ReturnUrl"] = returnUrl;
-                ViewData["LoginProvider"] = info.LoginProvider;
-                //var email = info.Principal.FindFirstValue(ClaimTypes.Email);
-                return View("ExternalLoginConfirmation", new ExternalLoginConfirmationViewModel { Email = email });
+                ModelState.AddModelError(string.Empty, _sharedLocalizer["INVALID_LOGIN_ATTEMPT"]);
+                return View(await BuildLoginViewModelAsync(model));
             }
         }
 
-        //
-        // POST: /Account/ExternalLoginConfirmation
-        [HttpPost]
-        [AllowAnonymous]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> ExternalLoginConfirmation(ExternalLoginConfirmationViewModel model, string returnUrl = null, string loginProvider = null)
+        // If we got this far, something failed, redisplay form
+        return View(await BuildLoginViewModelAsync(model));
+    }
+
+    [HttpGet]
+    [AllowAnonymous]
+    public async Task<IActionResult> LoginFido2Mfa(string provider, bool rememberMe, string returnUrl = null)
+    {
+        // Require that the user has already logged in via username/password or external login
+        var user = await _signInManager.GetTwoFactorAuthenticationUserAsync();
+        if (user == null)
         {
-            if (ModelState.IsValid)
+            return View("Error");
+        }
+
+        if (string.IsNullOrEmpty(provider))
+        {
+            provider = "fido2";
+        }
+
+        return View(new MfaModel { /*Provider = provider,*/ ReturnUrl = returnUrl, RememberMe = rememberMe });
+    }
+
+    [HttpGet]
+    [AllowAnonymous]
+    public IActionResult ErrorEnable2FA()
+    {
+        return View();
+    }
+
+    [HttpGet]
+    public async Task<IActionResult> Logout(string logoutId)
+    {
+        // build a model so the logout page knows what to display
+        var vm = await BuildLogoutViewModelAsync(logoutId);
+
+        if (vm.ShowLogoutPrompt == false)
+        {
+            // if the request for logout was properly authenticated from IdentityServer, then
+            // we don't need to show the prompt and can just log the user out directly.
+            return await Logout(vm);
+        }
+
+        return View(vm);
+    }
+
+    /// <summary>
+    /// Handle logout page postback
+    /// </summary>
+    [HttpPost]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> Logout(LogoutViewModel model)
+    {
+        var idp = User?.FindFirst(JwtClaimTypes.IdentityProvider)?.Value;
+        var subjectId = HttpContext.User.Identity.GetSubjectId();
+
+        if (idp != null && idp != IdentityServerConstants.LocalIdentityProvider)
+        {
+            if (model.LogoutId == null)
             {
-                // Get the information about the user from the external login provider
-                var info = await _signInManager.GetExternalLoginInfoAsync();
-                if (info == null)
-                {
-                    return View("ExternalLoginFailure");
-                }
-                var user = new ApplicationUser { UserName = model.Email, Email = model.Email };
-                var result = await _userManager.CreateAsync(user);
+                // if there's no current logout context, we need to create one
+                // this captures necessary info from the current logged in user
+                // before we signout and redirect away to the external IdP for signout
+                model.LogoutId = await _interaction.CreateLogoutContextAsync();
+            }
+
+            // string url = "/Account/Logout?logoutId=" + model.LogoutId;
+            try
+            {
+                await _signInManager.SignOutAsync();
+                // await HttpContext.Authentication.SignOutAsync(idp, new AuthenticationProperties { RedirectUri = url });
+            }
+            catch (NotSupportedException)
+            {
+            }
+        }
+
+        // delete authentication cookie
+        await _signInManager.SignOutAsync();
+
+        // set this so UI rendering sees an anonymous user
+        HttpContext.User = new ClaimsPrincipal(new ClaimsIdentity());
+
+        // get context information (client name, post logout redirect URI and iframe for federated signout)
+        var logout = await _interaction.GetLogoutContextAsync(model.LogoutId);
+
+        var vm = new LoggedOutViewModel
+        {
+            AutomaticRedirectAfterSignOut = AccountOptions.AutomaticRedirectAfterSignOut,
+            PostLogoutRedirectUri = logout?.PostLogoutRedirectUri,
+            ClientName = string.IsNullOrEmpty(logout?.ClientName) ? logout?.ClientId : logout?.ClientName,
+            SignOutIframeUrl = logout?.SignOutIFrameUrl,
+            LogoutId = model.LogoutId
+        };
+
+
+        await _persistedGrantService.RemoveAllGrantsAsync(subjectId, "angular2client");
+
+        return View("LoggedOut", vm);
+    }
+
+    //
+    // GET: /Account/Register
+    [HttpGet]
+    [AllowAnonymous]
+    public IActionResult Register(string returnUrl = null)
+    {
+        ViewData["ReturnUrl"] = returnUrl;
+        return View();
+    }
+
+    //
+    // POST: /Account/Register
+    [HttpPost]
+    [AllowAnonymous]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> Register(RegisterViewModel model, string returnUrl = null)
+    {
+        ViewData["ReturnUrl"] = returnUrl;
+        if (ModelState.IsValid)
+        {
+            var user = new ApplicationUser
+            {
+                UserName = model.Email,
+                Email = model.Email,
+                IsAdmin = false
+            };
+            var result = await _userManager.CreateAsync(user, model.Password);
+            if (result.Succeeded)
+            {
+                //var code = await _userManager.GenerateEmailConfirmationTokenAsync(user);
+                // WebEncoders.Base64UrlEncode(Encoding.UTF8.GetBytes(code));
+                //var callbackUrl = Url.Action("ConfirmEmail", "Account", new { userId = user.Id, code = code }, protocol: HttpContext.Request.Scheme);
+                //await _emailSender.SendEmailAsync(model.Email, "Confirm your account",
+                //    $"Please confirm your account by clicking this link: <a href='{HtmlEncoder.Default.Encode(callbackUrl)}'>link</a>");
+                //await _signInManager.SignInAsync(user, isPersistent: false);
+                //_logger.LogInformation(3, "User created a new account with password.");
+                return RedirectToLocal(returnUrl);
+            }
+            AddErrors(result);
+        }
+
+        // If we got this far, something failed, redisplay form
+        return View(model);
+    }
+
+    //
+    // POST: /Account/ExternalLogin
+    //[ValidateAntiForgeryToken]
+    [HttpPost]
+    [HttpGet]
+    [AllowAnonymous]
+    public IActionResult ExternalLogin(string provider, string returnUrl = null)
+    {
+        // Request a redirect to the external login provider.
+        var redirectUrl = Url.Action("ExternalLoginCallback", "Account", new { ReturnUrl = returnUrl });
+        var properties = _signInManager.ConfigureExternalAuthenticationProperties(provider, redirectUrl);
+        return Challenge(properties, provider);
+    }
+
+    //
+    // GET: /Account/ExternalLoginCallback
+    [HttpGet]
+    [AllowAnonymous]
+    public async Task<IActionResult> ExternalLoginCallback(string returnUrl = null, string remoteError = null)
+    {
+        var context = await _interaction.GetAuthorizationContextAsync(returnUrl);
+        var requires2Fa = context?.AcrValues.Count(t => t.Contains("mfa")) >= 1;
+
+        if (remoteError != null)
+        {
+            ModelState.AddModelError(string.Empty, _sharedLocalizer["EXTERNAL_PROVIDER_ERROR", remoteError]);
+            return View(nameof(Login));
+        }
+        var info = await _signInManager.GetExternalLoginInfoAsync();
+
+        if (info == null)
+        {
+            return RedirectToAction(nameof(Login));
+        }
+
+        var email = info.Principal.FindFirstValue(ClaimTypes.Email);
+
+        if (!string.IsNullOrEmpty(email))
+        {
+            var user = await _userManager.FindByNameAsync(email);
+            if (user != null && !user.TwoFactorEnabled && requires2Fa)
+            {
+                return RedirectToAction(nameof(ErrorEnable2FA));
+            }
+        }
+
+        // Sign in the user with this external login provider if the user already has a login.
+        var result = await _signInManager.ExternalLoginSignInAsync(info.LoginProvider, info.ProviderKey, isPersistent: false);
+        if (result.Succeeded)
+        {
+            _logger.LogInformation(5, "User logged in with {Name} provider.", info.LoginProvider);
+            return RedirectToLocal(returnUrl);
+        }
+        if (result.RequiresTwoFactor)
+        {
+            var fido2ItemExistsForUser = await _fido2Storage.GetCredentialsByUserNameAsync(email);
+            if (fido2ItemExistsForUser.Count > 0)
+            {
+                return RedirectToAction(nameof(LoginFido2Mfa), new { ReturnUrl = returnUrl });
+            }
+            else
+            {
+                return RedirectToAction(nameof(SendCode), new { ReturnUrl = returnUrl });
+            }
+        }
+        if (result.IsLockedOut)
+        {
+            return View("Lockout");
+        }
+        else
+        {
+            // If the user does not have an account, then ask the user to create an account.
+            ViewData["ReturnUrl"] = returnUrl;
+            ViewData["LoginProvider"] = info.LoginProvider;
+            //var email = info.Principal.FindFirstValue(ClaimTypes.Email);
+            return View("ExternalLoginConfirmation", new ExternalLoginConfirmationViewModel { Email = email });
+        }
+    }
+
+    //
+    // POST: /Account/ExternalLoginConfirmation
+    [HttpPost]
+    [AllowAnonymous]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> ExternalLoginConfirmation(ExternalLoginConfirmationViewModel model, string returnUrl = null, string loginProvider = null)
+    {
+        if (ModelState.IsValid)
+        {
+            // Get the information about the user from the external login provider
+            var info = await _signInManager.GetExternalLoginInfoAsync();
+            if (info == null)
+            {
+                return View("ExternalLoginFailure");
+            }
+            var user = new ApplicationUser { UserName = model.Email, Email = model.Email };
+            var result = await _userManager.CreateAsync(user);
+            if (result.Succeeded)
+            {
+                result = await _userManager.AddLoginAsync(user, info);
                 if (result.Succeeded)
                 {
-                    result = await _userManager.AddLoginAsync(user, info);
-                    if (result.Succeeded)
-                    {
-                        await _signInManager.SignInAsync(user, isPersistent: false);
-                        _logger.LogInformation(6, "User created an account using {Name} provider.", info.LoginProvider);
-                        return RedirectToLocal(returnUrl);
-                    }
+                    await _signInManager.SignInAsync(user, isPersistent: false);
+                    _logger.LogInformation(6, "User created an account using {Name} provider.", info.LoginProvider);
+                    return RedirectToLocal(returnUrl);
                 }
-                AddErrors(result);
             }
-
-            ViewData["ReturnUrl"] = returnUrl;
-            ViewData["LoginProvider"] = loginProvider;
-            return View(model);
+            AddErrors(result);
         }
 
-        // GET: /Account/ConfirmEmail
-        [HttpGet]
-        [AllowAnonymous]
-        public async Task<IActionResult> ConfirmEmail(string userId, string code)
+        ViewData["ReturnUrl"] = returnUrl;
+        ViewData["LoginProvider"] = loginProvider;
+        return View(model);
+    }
+
+    // GET: /Account/ConfirmEmail
+    [HttpGet]
+    [AllowAnonymous]
+    public async Task<IActionResult> ConfirmEmail(string userId, string code)
+    {
+        if (userId == null || code == null)
         {
-            if (userId == null || code == null)
-            {
-                return View("Error");
-            }
-            var user = await _userManager.FindByIdAsync(userId);
+            return View("Error");
+        }
+        var user = await _userManager.FindByIdAsync(userId);
+        if (user == null)
+        {
+            return View("Error");
+        }
+        code = Encoding.UTF8.GetString(WebEncoders.Base64UrlDecode(code));
+        var result = await _userManager.ConfirmEmailAsync(user, code);
+        return View(result.Succeeded ? "ConfirmEmail" : "Error");
+    }
+
+    //
+    // GET: /Account/ForgotPassword
+    [HttpGet]
+    [AllowAnonymous]
+    public IActionResult ForgotPassword()
+    {
+        return View();
+    }
+
+    //
+    // POST: /Account/ForgotPassword
+    [HttpPost]
+    [AllowAnonymous]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> ForgotPassword(ForgotPasswordViewModel model)
+    {
+        if (ModelState.IsValid)
+        {
+            var user = await _userManager.FindByNameAsync(model.Email);
+            // TODO add this is all users need to be validated
+            // if (user == null || !(await _userManager.IsEmailConfirmedAsync(user)))
             if (user == null)
             {
-                return View("Error");
-            }
-            code = Encoding.UTF8.GetString(WebEncoders.Base64UrlDecode(code));
-            var result = await _userManager.ConfirmEmailAsync(user, code);
-            return View(result.Succeeded ? "ConfirmEmail" : "Error");
-        }
-
-        //
-        // GET: /Account/ForgotPassword
-        [HttpGet]
-        [AllowAnonymous]
-        public IActionResult ForgotPassword()
-        {
-            return View();
-        }
-
-        //
-        // POST: /Account/ForgotPassword
-        [HttpPost]
-        [AllowAnonymous]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> ForgotPassword(ForgotPasswordViewModel model)
-        {
-            if (ModelState.IsValid)
-            {
-                var user = await _userManager.FindByNameAsync(model.Email);
-                // TODO add this is all users need to be validated
-                // if (user == null || !(await _userManager.IsEmailConfirmedAsync(user)))
-                if (user == null)
-                {
-                    // Don't reveal that the user does not exist or is not confirmed
-                    return View("ForgotPasswordConfirmation");
-                }
-
-                // For more information on how to enable account confirmation and password reset please visit http://go.microsoft.com/fwlink/?LinkID=532713
-                // Send an email with this link
-                var code = await _userManager.GeneratePasswordResetTokenAsync(user);
-                code = WebEncoders.Base64UrlEncode(Encoding.UTF8.GetBytes(code));
-                var callbackUrl = Url.Action("ResetPassword", "Account", new { userId = user.Id, code = code }, protocol: HttpContext.Request.Scheme);
-                await _emailSender.SendEmail(
-                   model.Email,
-                   "Reset Password",
-                   $"Please reset your password by clicking here: {HtmlEncoder.Default.Encode(callbackUrl)}",
-                   "Hi Sir");
-
+                // Don't reveal that the user does not exist or is not confirmed
                 return View("ForgotPasswordConfirmation");
             }
 
-            // If we got this far, something failed, redisplay form
+            // For more information on how to enable account confirmation and password reset please visit http://go.microsoft.com/fwlink/?LinkID=532713
+            // Send an email with this link
+            var code = await _userManager.GeneratePasswordResetTokenAsync(user);
+            code = WebEncoders.Base64UrlEncode(Encoding.UTF8.GetBytes(code));
+            var callbackUrl = Url.Action("ResetPassword", "Account", new { userId = user.Id, code = code }, protocol: HttpContext.Request.Scheme);
+            await _emailSender.SendEmail(
+               model.Email,
+               "Reset Password",
+               $"Please reset your password by clicking here: {HtmlEncoder.Default.Encode(callbackUrl)}",
+               "Hi Sir");
+
+            return View("ForgotPasswordConfirmation");
+        }
+
+        // If we got this far, something failed, redisplay form
+        return View(model);
+    }
+
+    //
+    // GET: /Account/ForgotPasswordConfirmation
+    [HttpGet]
+    [AllowAnonymous]
+    public IActionResult ForgotPasswordConfirmation()
+    {
+        return View();
+    }
+
+    //
+    // GET: /Account/ResetPassword
+    [HttpGet]
+    [AllowAnonymous]
+    public IActionResult ResetPassword(string code = null)
+    {
+        return code == null ? View("Error") : View();
+    }
+
+    //
+    // POST: /Account/ResetPassword
+    [HttpPost]
+    [AllowAnonymous]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> ResetPassword(ResetPasswordViewModel model)
+    {
+        if (!ModelState.IsValid)
+        {
             return View(model);
         }
+        var user = await _userManager.FindByNameAsync(model.Email);
+        if (user == null)
+        {
+            // Don't reveal that the user does not exist
+            return RedirectToAction(nameof(AccountController.ResetPasswordConfirmation), "Account");
+        }
+        var code = Encoding.UTF8.GetString(WebEncoders.Base64UrlDecode(model.Code));
+        var result = await _userManager.ResetPasswordAsync(user, code, model.Password);
+        if (result.Succeeded)
+        {
+            return RedirectToAction(nameof(AccountController.ResetPasswordConfirmation), "Account");
+        }
+        AddErrors(result);
+        return View();
+    }
 
-        //
-        // GET: /Account/ForgotPasswordConfirmation
-        [HttpGet]
-        [AllowAnonymous]
-        public IActionResult ForgotPasswordConfirmation()
+    //
+    // GET: /Account/ResetPasswordConfirmation
+    [HttpGet]
+    [AllowAnonymous]
+    public IActionResult ResetPasswordConfirmation()
+    {
+        return View();
+    }
+
+    //
+    // GET: /Account/SendCode
+    [HttpGet]
+    [AllowAnonymous]
+    public async Task<ActionResult> SendCode(string returnUrl = null, bool rememberMe = false)
+    {
+        var user = await _signInManager.GetTwoFactorAuthenticationUserAsync();
+        if (user == null)
+        {
+            return View("Error");
+        }
+        var userFactors = await _userManager.GetValidTwoFactorProvidersAsync(user);
+        var factorOptions = userFactors.Select(purpose => new SelectListItem { Text = purpose, Value = purpose }).ToList();
+        return View(new SendCodeViewModel { Providers = factorOptions, ReturnUrl = returnUrl, RememberMe = rememberMe });
+    }
+
+    //
+    // POST: /Account/SendCode
+    [HttpPost]
+    [AllowAnonymous]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> SendCode(SendCodeViewModel model)
+    {
+        if (!ModelState.IsValid)
         {
             return View();
         }
 
-        //
-        // GET: /Account/ResetPassword
-        [HttpGet]
-        [AllowAnonymous]
-        public IActionResult ResetPassword(string code = null)
+        var user = await _signInManager.GetTwoFactorAuthenticationUserAsync();
+        if (user == null)
         {
-            return code == null ? View("Error") : View();
+            return View("Error");
         }
-
-        //
-        // POST: /Account/ResetPassword
-        [HttpPost]
-        [AllowAnonymous]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> ResetPassword(ResetPasswordViewModel model)
+        if (model.SelectedProvider == "Authenticator")
         {
-            if (!ModelState.IsValid)
-            {
-                return View(model);
-            }
-            var user = await _userManager.FindByNameAsync(model.Email);
-            if (user == null)
-            {
-                // Don't reveal that the user does not exist
-                return RedirectToAction(nameof(AccountController.ResetPasswordConfirmation), "Account");
-            }
-            var code = Encoding.UTF8.GetString(WebEncoders.Base64UrlDecode(model.Code));
-            var result = await _userManager.ResetPasswordAsync(user, code, model.Password);
-            if (result.Succeeded)
-            {
-                return RedirectToAction(nameof(AccountController.ResetPasswordConfirmation), "Account");
-            }
-            AddErrors(result);
-            return View();
-        }
-
-        //
-        // GET: /Account/ResetPasswordConfirmation
-        [HttpGet]
-        [AllowAnonymous]
-        public IActionResult ResetPasswordConfirmation()
-        {
-            return View();
-        }
-
-        //
-        // GET: /Account/SendCode
-        [HttpGet]
-        [AllowAnonymous]
-        public async Task<ActionResult> SendCode(string returnUrl = null, bool rememberMe = false)
-        {
-            var user = await _signInManager.GetTwoFactorAuthenticationUserAsync();
-            if (user == null)
-            {
-                return View("Error");
-            }
-            var userFactors = await _userManager.GetValidTwoFactorProvidersAsync(user);
-            var factorOptions = userFactors.Select(purpose => new SelectListItem { Text = purpose, Value = purpose }).ToList();
-            return View(new SendCodeViewModel { Providers = factorOptions, ReturnUrl = returnUrl, RememberMe = rememberMe });
-        }
-
-        //
-        // POST: /Account/SendCode
-        [HttpPost]
-        [AllowAnonymous]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> SendCode(SendCodeViewModel model)
-        {
-            if (!ModelState.IsValid)
-            {
-                return View();
-            }
-
-            var user = await _signInManager.GetTwoFactorAuthenticationUserAsync();
-            if (user == null)
-            {
-                return View("Error");
-            }
-            if (model.SelectedProvider == "Authenticator")
-            {
-                return RedirectToAction(nameof(VerifyCode), new { Provider = model.SelectedProvider, ReturnUrl = model.ReturnUrl, RememberMe = model.RememberMe });
-            }
-
-            // Email used
-            // Generate the token and send it
-            var code = await _userManager.GenerateTwoFactorTokenAsync(user, model.SelectedProvider);
-            code = WebEncoders.Base64UrlEncode(Encoding.UTF8.GetBytes(code));
-            if (string.IsNullOrWhiteSpace(code))
-            {
-                return View("Error");
-            }
-
-            var message = "Your security code is: " + code;
-            if (model.SelectedProvider == "Email")
-            {
-                await _emailSender.SendEmail(await _userManager.GetEmailAsync(user), "Security Code", message, "Hi Sir");
-            }
-
             return RedirectToAction(nameof(VerifyCode), new { Provider = model.SelectedProvider, ReturnUrl = model.ReturnUrl, RememberMe = model.RememberMe });
         }
 
-        //
-        // GET: /Account/VerifyCode
-        [HttpGet]
-        [AllowAnonymous]
-        public async Task<IActionResult> VerifyCode(string provider, bool rememberMe, string returnUrl = null)
+        // Email used
+        // Generate the token and send it
+        var code = await _userManager.GenerateTwoFactorTokenAsync(user, model.SelectedProvider);
+        code = WebEncoders.Base64UrlEncode(Encoding.UTF8.GetBytes(code));
+        if (string.IsNullOrWhiteSpace(code))
         {
-            // Require that the user has already logged in via username/password or external login
-            var user = await _signInManager.GetTwoFactorAuthenticationUserAsync();
-            if (user == null)
-            {
-                return View("Error");
-            }
-
-            if (string.IsNullOrEmpty(provider))
-            {
-                provider = "Authenticator";
-            }
-
-            return View(new VerifyCodeViewModel { Provider = provider, ReturnUrl = returnUrl, RememberMe = rememberMe });
+            return View("Error");
         }
 
-        //
-        // POST: /Account/VerifyCode
-        [HttpPost]
-        [AllowAnonymous]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> VerifyCode(VerifyCodeViewModel model)
+        var message = "Your security code is: " + code;
+        if (model.SelectedProvider == "Email")
         {
-            if (!ModelState.IsValid)
-            {
-                return View(model);
-            }
-
-            if (string.IsNullOrEmpty(model.Provider))
-            {
-                model.Provider = "Authenticator";
-            }
-
-            // The following code protects for brute force attacks against the two factor codes.
-            // If a user enters incorrect codes for a specified amount of time then the user account
-            // will be locked out for a specified amount of time.
-            var result = await _signInManager.TwoFactorSignInAsync(model.Provider, model.Code, model.RememberMe, model.RememberBrowser);
-            if (result.Succeeded)
-            {
-                return RedirectToLocal(model.ReturnUrl);
-            }
-            if (result.IsLockedOut)
-            {
-                _logger.LogWarning(7, "User account locked out.");
-                return View("Lockout");
-            }
-            else
-            {
-                ModelState.AddModelError(string.Empty, _sharedLocalizer["INVALID_CODE"]);
-                return View(model);
-            }
+            await _emailSender.SendEmail(await _userManager.GetEmailAsync(user), "Security Code", message, "Hi Sir");
         }
 
-        private async Task<LoginViewModel> BuildLoginViewModelAsync(string returnUrl)
+        return RedirectToAction(nameof(VerifyCode), new { Provider = model.SelectedProvider, ReturnUrl = model.ReturnUrl, RememberMe = model.RememberMe });
+    }
+
+    //
+    // GET: /Account/VerifyCode
+    [HttpGet]
+    [AllowAnonymous]
+    public async Task<IActionResult> VerifyCode(string provider, bool rememberMe, string returnUrl = null)
+    {
+        // Require that the user has already logged in via username/password or external login
+        var user = await _signInManager.GetTwoFactorAuthenticationUserAsync();
+        if (user == null)
         {
-            var context = await _interaction.GetAuthorizationContextAsync(returnUrl);
-            if (context?.IdP != null && await _schemeProvider.GetSchemeAsync(context.IdP) != null)
+            return View("Error");
+        }
+
+        if (string.IsNullOrEmpty(provider))
+        {
+            provider = "Authenticator";
+        }
+
+        return View(new VerifyCodeViewModel { Provider = provider, ReturnUrl = returnUrl, RememberMe = rememberMe });
+    }
+
+    //
+    // POST: /Account/VerifyCode
+    [HttpPost]
+    [AllowAnonymous]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> VerifyCode(VerifyCodeViewModel model)
+    {
+        if (!ModelState.IsValid)
+        {
+            return View(model);
+        }
+
+        if (string.IsNullOrEmpty(model.Provider))
+        {
+            model.Provider = "Authenticator";
+        }
+
+        // The following code protects for brute force attacks against the two factor codes.
+        // If a user enters incorrect codes for a specified amount of time then the user account
+        // will be locked out for a specified amount of time.
+        var result = await _signInManager.TwoFactorSignInAsync(model.Provider, model.Code, model.RememberMe, model.RememberBrowser);
+        if (result.Succeeded)
+        {
+            return RedirectToLocal(model.ReturnUrl);
+        }
+        if (result.IsLockedOut)
+        {
+            _logger.LogWarning(7, "User account locked out.");
+            return View("Lockout");
+        }
+        else
+        {
+            ModelState.AddModelError(string.Empty, _sharedLocalizer["INVALID_CODE"]);
+            return View(model);
+        }
+    }
+
+    private async Task<LoginViewModel> BuildLoginViewModelAsync(string returnUrl)
+    {
+        var context = await _interaction.GetAuthorizationContextAsync(returnUrl);
+        if (context?.IdP != null && await _schemeProvider.GetSchemeAsync(context.IdP) != null)
+        {
+            var local = context.IdP == IdentityServer4.IdentityServerConstants.LocalIdentityProvider;
+
+            // this is meant to short circuit the UI and only trigger the one external IdP
+            var vm = new LoginViewModel
             {
-                var local = context.IdP == IdentityServer4.IdentityServerConstants.LocalIdentityProvider;
-
-                // this is meant to short circuit the UI and only trigger the one external IdP
-                var vm = new LoginViewModel
-                {
-                    EnableLocalLogin = local,
-                    ReturnUrl = returnUrl,
-                    Email = context?.LoginHint,
-                };
-
-                if (!local)
-                {
-                    vm.ExternalProviders = new[] { new ExternalProvider { AuthenticationScheme = context.IdP } };
-                }
-
-                return vm;
-            }
-
-            var schemes = await _schemeProvider.GetAllSchemesAsync();
-
-            var providers = schemes
-                .Where(x => x.DisplayName != null)
-                .Select(x => new ExternalProvider
-                {
-                    DisplayName = x.DisplayName ?? x.Name,
-                    AuthenticationScheme = x.Name
-                }).ToList();
-
-            var allowLocal = true;
-            if (context?.Client.ClientId != null)
-            {
-                var client = await _clientStore.FindEnabledClientByIdAsync(context.Client.ClientId);
-                if (client != null)
-                {
-                    allowLocal = client.EnableLocalLogin;
-
-                    if (client.IdentityProviderRestrictions != null && client.IdentityProviderRestrictions.Any())
-                    {
-                        providers = providers.Where(provider => client.IdentityProviderRestrictions.Contains(provider.AuthenticationScheme)).ToList();
-                    }
-                }
-            }
-
-            return new LoginViewModel
-            {
-                AllowRememberLogin = AccountOptions.AllowRememberLogin,
-                EnableLocalLogin = allowLocal && AccountOptions.AllowLocalLogin,
+                EnableLocalLogin = local,
                 ReturnUrl = returnUrl,
                 Email = context?.LoginHint,
-                ExternalProviders = providers.ToArray()
             };
-        }
 
-        private async Task<LoginViewModel> BuildLoginViewModelAsync(LoginInputModel model)
-        {
-            var vm = await BuildLoginViewModelAsync(model.ReturnUrl);
-            vm.Email = model.Email;
-            vm.RememberLogin = model.RememberLogin;
+            if (!local)
+            {
+                vm.ExternalProviders = new[] { new ExternalProvider { AuthenticationScheme = context.IdP } };
+            }
+
             return vm;
         }
 
-        private async Task<LogoutViewModel> BuildLogoutViewModelAsync(string logoutId)
+        var schemes = await _schemeProvider.GetAllSchemesAsync();
+
+        var providers = schemes
+            .Where(x => x.DisplayName != null)
+            .Select(x => new ExternalProvider
+            {
+                DisplayName = x.DisplayName ?? x.Name,
+                AuthenticationScheme = x.Name
+            }).ToList();
+
+        var allowLocal = true;
+        if (context?.Client.ClientId != null)
         {
-            var vm = new LogoutViewModel { LogoutId = logoutId, ShowLogoutPrompt = AccountOptions.ShowLogoutPrompt };
-
-            if (User?.Identity.IsAuthenticated != true)
+            var client = await _clientStore.FindEnabledClientByIdAsync(context.Client.ClientId);
+            if (client != null)
             {
-                // if the user is not authenticated, then just show logged out page
-                vm.ShowLogoutPrompt = false;
-                return vm;
-            }
+                allowLocal = client.EnableLocalLogin;
 
-            var context = await _interaction.GetLogoutContextAsync(logoutId);
-            if (context?.ShowSignoutPrompt == false)
-            {
-                // it's safe to automatically sign-out
-                vm.ShowLogoutPrompt = false;
-                return vm;
-            }
-
-            // show the logout prompt. this prevents attacks where the user
-            // is automatically signed out by another malicious web page.
-            return vm;
-        }
-
-        private async Task<LoggedOutViewModel> BuildLoggedOutViewModelAsync(string logoutId)
-        {
-            // get context information (client name, post logout redirect URI and iframe for federated signout)
-            var logout = await _interaction.GetLogoutContextAsync(logoutId);
-
-            var vm = new LoggedOutViewModel
-            {
-                AutomaticRedirectAfterSignOut = AccountOptions.AutomaticRedirectAfterSignOut,
-                PostLogoutRedirectUri = logout?.PostLogoutRedirectUri,
-                ClientName = string.IsNullOrEmpty(logout?.ClientName) ? logout?.ClientId : logout?.ClientName,
-                SignOutIframeUrl = logout?.SignOutIFrameUrl,
-                LogoutId = logoutId
-            };
-
-            if (User?.Identity.IsAuthenticated == true)
-            {
-                var idp = User.FindFirst(JwtClaimTypes.IdentityProvider)?.Value;
-                if (idp != null && idp != IdentityServer4.IdentityServerConstants.LocalIdentityProvider)
+                if (client.IdentityProviderRestrictions != null && client.IdentityProviderRestrictions.Any())
                 {
-                    var providerSupportsSignout = await HttpContext.GetSchemeSupportsSignOutAsync(idp);
-                    if (providerSupportsSignout)
-                    {
-                        if (vm.LogoutId == null)
-                        {
-                            // if there's no current logout context, we need to create one
-                            // this captures necessary info from the current logged in user
-                            // before we signout and redirect away to the external IdP for signout
-                            vm.LogoutId = await _interaction.CreateLogoutContextAsync();
-                        }
-
-                        vm.ExternalAuthenticationScheme = idp;
-                    }
+                    providers = providers.Where(provider => client.IdentityProviderRestrictions.Contains(provider.AuthenticationScheme)).ToList();
                 }
             }
+        }
 
+        return new LoginViewModel
+        {
+            AllowRememberLogin = AccountOptions.AllowRememberLogin,
+            EnableLocalLogin = allowLocal && AccountOptions.AllowLocalLogin,
+            ReturnUrl = returnUrl,
+            Email = context?.LoginHint,
+            ExternalProviders = providers.ToArray()
+        };
+    }
+
+    private async Task<LoginViewModel> BuildLoginViewModelAsync(LoginInputModel model)
+    {
+        var vm = await BuildLoginViewModelAsync(model.ReturnUrl);
+        vm.Email = model.Email;
+        vm.RememberLogin = model.RememberLogin;
+        return vm;
+    }
+
+    private async Task<LogoutViewModel> BuildLogoutViewModelAsync(string logoutId)
+    {
+        var vm = new LogoutViewModel { LogoutId = logoutId, ShowLogoutPrompt = AccountOptions.ShowLogoutPrompt };
+
+        if (User?.Identity.IsAuthenticated != true)
+        {
+            // if the user is not authenticated, then just show logged out page
+            vm.ShowLogoutPrompt = false;
             return vm;
         }
-        #region Helpers
 
-        private void AddErrors(IdentityResult result)
+        var context = await _interaction.GetLogoutContextAsync(logoutId);
+        if (context?.ShowSignoutPrompt == false)
         {
-            foreach (var error in result.Errors)
-            {
-                ModelState.AddModelError(string.Empty, error.Description);
-            }
+            // it's safe to automatically sign-out
+            vm.ShowLogoutPrompt = false;
+            return vm;
         }
 
-        private IActionResult RedirectToLocal(string returnUrl)
-        {
-            if (Url.IsLocalUrl(returnUrl))
-            {
-                return Redirect(returnUrl);
-            }
-            else
-            {
-                return RedirectToAction(nameof(HomeController.Index), "Home");
-            }
-        }
-
-        #endregion
+        // show the logout prompt. this prevents attacks where the user
+        // is automatically signed out by another malicious web page.
+        return vm;
     }
+
+    private async Task<LoggedOutViewModel> BuildLoggedOutViewModelAsync(string logoutId)
+    {
+        // get context information (client name, post logout redirect URI and iframe for federated signout)
+        var logout = await _interaction.GetLogoutContextAsync(logoutId);
+
+        var vm = new LoggedOutViewModel
+        {
+            AutomaticRedirectAfterSignOut = AccountOptions.AutomaticRedirectAfterSignOut,
+            PostLogoutRedirectUri = logout?.PostLogoutRedirectUri,
+            ClientName = string.IsNullOrEmpty(logout?.ClientName) ? logout?.ClientId : logout?.ClientName,
+            SignOutIframeUrl = logout?.SignOutIFrameUrl,
+            LogoutId = logoutId
+        };
+
+        if (User?.Identity.IsAuthenticated == true)
+        {
+            var idp = User.FindFirst(JwtClaimTypes.IdentityProvider)?.Value;
+            if (idp != null && idp != IdentityServer4.IdentityServerConstants.LocalIdentityProvider)
+            {
+                var providerSupportsSignout = await HttpContext.GetSchemeSupportsSignOutAsync(idp);
+                if (providerSupportsSignout)
+                {
+                    if (vm.LogoutId == null)
+                    {
+                        // if there's no current logout context, we need to create one
+                        // this captures necessary info from the current logged in user
+                        // before we signout and redirect away to the external IdP for signout
+                        vm.LogoutId = await _interaction.CreateLogoutContextAsync();
+                    }
+
+                    vm.ExternalAuthenticationScheme = idp;
+                }
+            }
+        }
+
+        return vm;
+    }
+    #region Helpers
+
+    private void AddErrors(IdentityResult result)
+    {
+        foreach (var error in result.Errors)
+        {
+            ModelState.AddModelError(string.Empty, error.Description);
+        }
+    }
+
+    private IActionResult RedirectToLocal(string returnUrl)
+    {
+        if (Url.IsLocalUrl(returnUrl))
+        {
+            return Redirect(returnUrl);
+        }
+        else
+        {
+            return RedirectToAction(nameof(HomeController.Index), "Home");
+        }
+    }
+
+    #endregion
 }
