@@ -4,266 +4,260 @@ using IdentityServer4.Services;
 using IdentityServer4.Extensions;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-using System.Linq;
-using System.Threading.Tasks;
 using IdentityServer4.Validation;
-using System.Collections.Generic;
-using System;
 using StsServerIdentity.Filters;
 using StsServerIdentity.Models;
 using Microsoft.Extensions.Localization;
 using StsServerIdentity.Resources;
 using System.Reflection;
 
-namespace StsServerIdentity.Controllers
+namespace StsServerIdentity.Controllers;
+
+/// <summary>
+/// This controller processes the consent UI
+/// </summary>
+[SecurityHeaders]
+[Authorize]
+public class ConsentController : Controller
 {
-    /// <summary>
-    /// This controller processes the consent UI
-    /// </summary>
-    [SecurityHeaders]
-    [Authorize]
-    public class ConsentController : Controller
+    private readonly IIdentityServerInteractionService _interaction;
+    private readonly IEventService _events;
+    private readonly ILogger<ConsentController> _logger;
+    private readonly IStringLocalizer _sharedLocalizer;
+
+    public ConsentController(
+        IIdentityServerInteractionService interaction,
+        IEventService events,
+        ILogger<ConsentController> logger,
+        IStringLocalizerFactory factory)
     {
-        private readonly IIdentityServerInteractionService _interaction;
-        private readonly IEventService _events;
-        private readonly ILogger<ConsentController> _logger;
-        private readonly IStringLocalizer _sharedLocalizer;
+        _interaction = interaction;
+        _events = events;
+        _logger = logger;
 
-        public ConsentController(
-            IIdentityServerInteractionService interaction,
-            IEventService events,
-            ILogger<ConsentController> logger,
-            IStringLocalizerFactory factory)
+        var type = typeof(SharedResource);
+        var assemblyName = new AssemblyName(type.GetTypeInfo().Assembly.FullName);
+        _sharedLocalizer = factory.Create("SharedResource", assemblyName.Name);
+    }
+
+    /// <summary>
+    /// Shows the consent screen
+    /// </summary>
+    /// <param name="returnUrl"></param>
+    /// <returns></returns>
+    [HttpGet]
+    public async Task<IActionResult> Index(string returnUrl)
+    {
+        var vm = await BuildViewModelAsync(returnUrl);
+        if (vm != null)
         {
-            _interaction = interaction;
-            _events = events;
-            _logger = logger;
-
-            var type = typeof(SharedResource);
-            var assemblyName = new AssemblyName(type.GetTypeInfo().Assembly.FullName);
-            _sharedLocalizer = factory.Create("SharedResource", assemblyName.Name);
+            return View("Index", vm);
         }
 
-        /// <summary>
-        /// Shows the consent screen
-        /// </summary>
-        /// <param name="returnUrl"></param>
-        /// <returns></returns>
-        [HttpGet]
-        public async Task<IActionResult> Index(string returnUrl)
+        return View("Error");
+    }
+
+    /// <summary>
+    /// Handles the consent screen postback
+    /// </summary>
+    [HttpPost]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> Index(ConsentInputModel model)
+    {
+        var result = await ProcessConsent(model);
+
+        if (result.IsRedirect)
         {
-            var vm = await BuildViewModelAsync(returnUrl);
-            if (vm != null)
+            var context = await _interaction.GetAuthorizationContextAsync(model.ReturnUrl);
+            if (context?.IsNativeClient() == true)
             {
-                return View("Index", vm);
+                // The client is native, so this change in how to
+                // return the response is for better UX for the end user.
+                return this.LoadingPage("Redirect", result.RedirectUri);
             }
 
-            return View("Error");
+            return Redirect(result.RedirectUri);
         }
 
-        /// <summary>
-        /// Handles the consent screen postback
-        /// </summary>
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Index(ConsentInputModel model)
+        if (result.HasValidationError)
         {
-            var result = await ProcessConsent(model);
+            ModelState.AddModelError(string.Empty, result.ValidationError);
+        }
 
-            if (result.IsRedirect)
+        if (result.ShowView)
+        {
+            return View("Index", result.ViewModel);
+        }
+
+        return View("Error");
+    }
+
+    /*****************************************/
+    /* helper APIs for the ConsentController */
+    /*****************************************/
+    private async Task<ProcessConsentResult> ProcessConsent(ConsentInputModel model)
+    {
+        var result = new ProcessConsentResult();
+
+        // validate return url is still valid
+        var request = await _interaction.GetAuthorizationContextAsync(model.ReturnUrl);
+        if (request == null) return result;
+
+        ConsentResponse grantedConsent = null;
+
+        // user clicked 'no' - send back the standard 'access_denied' response
+        if (model?.Button == "no")
+        {
+            grantedConsent = new ConsentResponse { Error = AuthorizationError.AccessDenied };
+
+            // emit event
+            await _events.RaiseAsync(new ConsentDeniedEvent(User.GetSubjectId(), request.Client.ClientId, request.ValidatedResources.RawScopeValues));
+        }
+        // user clicked 'yes' - validate the data
+        else if (model?.Button == "yes")
+        {
+            // if the user consented to some scope, build the response model
+            if (model.ScopesConsented != null && model.ScopesConsented.Any())
             {
-                var context = await _interaction.GetAuthorizationContextAsync(model.ReturnUrl);
-                if (context?.IsNativeClient() == true)
+                var scopes = model.ScopesConsented;
+                if (ConsentOptions.EnableOfflineAccess == false)
                 {
-                    // The client is native, so this change in how to
-                    // return the response is for better UX for the end user.
-                    return this.LoadingPage("Redirect", result.RedirectUri);
+                    scopes = scopes.Where(x => x != IdentityServer4.IdentityServerConstants.StandardScopes.OfflineAccess);
                 }
 
-                return Redirect(result.RedirectUri);
-            }
-
-            if (result.HasValidationError)
-            {
-                ModelState.AddModelError(string.Empty, result.ValidationError);
-            }
-
-            if (result.ShowView)
-            {
-                return View("Index", result.ViewModel);
-            }
-
-            return View("Error");
-        }
-
-        /*****************************************/
-        /* helper APIs for the ConsentController */
-        /*****************************************/
-        private async Task<ProcessConsentResult> ProcessConsent(ConsentInputModel model)
-        {
-            var result = new ProcessConsentResult();
-
-            // validate return url is still valid
-            var request = await _interaction.GetAuthorizationContextAsync(model.ReturnUrl);
-            if (request == null) return result;
-
-            ConsentResponse grantedConsent = null;
-
-            // user clicked 'no' - send back the standard 'access_denied' response
-            if (model?.Button == "no")
-            {
-                grantedConsent = new ConsentResponse { Error = AuthorizationError.AccessDenied };
+                grantedConsent = new ConsentResponse
+                {
+                    RememberConsent = model.RememberConsent,
+                    ScopesValuesConsented = scopes.ToArray(),
+                    Description = model.Description
+                };
 
                 // emit event
-                await _events.RaiseAsync(new ConsentDeniedEvent(User.GetSubjectId(), request.Client.ClientId, request.ValidatedResources.RawScopeValues));
-            }
-            // user clicked 'yes' - validate the data
-            else if (model?.Button == "yes")
-            {
-                // if the user consented to some scope, build the response model
-                if (model.ScopesConsented != null && model.ScopesConsented.Any())
-                {
-                    var scopes = model.ScopesConsented;
-                    if (ConsentOptions.EnableOfflineAccess == false)
-                    {
-                        scopes = scopes.Where(x => x != IdentityServer4.IdentityServerConstants.StandardScopes.OfflineAccess);
-                    }
-
-                    grantedConsent = new ConsentResponse
-                    {
-                        RememberConsent = model.RememberConsent,
-                        ScopesValuesConsented = scopes.ToArray(),
-                        Description = model.Description
-                    };
-
-                    // emit event
-                    await _events.RaiseAsync(new ConsentGrantedEvent(User.GetSubjectId(), request.Client.ClientId, request.ValidatedResources.RawScopeValues, grantedConsent.ScopesValuesConsented, grantedConsent.RememberConsent));
-                }
-                else
-                {
-                    result.ValidationError = _sharedLocalizer["You must pick at least one permission"];
-                }
+                await _events.RaiseAsync(new ConsentGrantedEvent(User.GetSubjectId(), request.Client.ClientId, request.ValidatedResources.RawScopeValues, grantedConsent.ScopesValuesConsented, grantedConsent.RememberConsent));
             }
             else
             {
-                result.ValidationError = _sharedLocalizer["Invalid selection"];
+                result.ValidationError = _sharedLocalizer["You must pick at least one permission"];
             }
-
-            if (grantedConsent != null)
-            {
-                // communicate outcome of consent back to identityserver
-                await _interaction.GrantConsentAsync(request, grantedConsent);
-
-                // indicate that's it ok to redirect back to authorization endpoint
-                result.RedirectUri = model.ReturnUrl;
-                result.Client = request.Client;
-            }
-            else
-            {
-                // we need to redisplay the consent UI
-                result.ViewModel = await BuildViewModelAsync(model.ReturnUrl, model);
-            }
-
-            return result;
         }
-
-        private async Task<ConsentViewModel> BuildViewModelAsync(string returnUrl, ConsentInputModel model = null)
+        else
         {
-            var request = await _interaction.GetAuthorizationContextAsync(returnUrl);
-            if (request != null)
-            {
-                return CreateConsentViewModel(model, returnUrl, request);
-            }
-            else
-            {
-                _logger.LogError("No consent request matching request: {0}", returnUrl);
-            }
-
-            return null;
+            result.ValidationError = _sharedLocalizer["Invalid selection"];
         }
 
-        private ConsentViewModel CreateConsentViewModel(
-            ConsentInputModel model, string returnUrl,
-            AuthorizationRequest request)
+        if (grantedConsent != null)
         {
-            var vm = new ConsentViewModel
-            {
-                RememberConsent = model?.RememberConsent ?? true,
-                ScopesConsented = model?.ScopesConsented ?? Enumerable.Empty<string>(),
-                Description = model?.Description,
+            // communicate outcome of consent back to identityserver
+            await _interaction.GrantConsentAsync(request, grantedConsent);
 
-                ReturnUrl = returnUrl,
-
-                ClientName = request.Client.ClientName ?? request.Client.ClientId,
-                ClientUrl = request.Client.ClientUri,
-                ClientLogoUrl = request.Client.LogoUri,
-                AllowRememberConsent = request.Client.AllowRememberConsent
-            };
-
-            vm.IdentityScopes = request.ValidatedResources.Resources.IdentityResources.Select(x => CreateScopeViewModel(x, vm.ScopesConsented.Contains(x.Name) || model == null)).ToArray();
-
-            var apiScopes = new List<ScopeViewModel>();
-            foreach (var parsedScope in request.ValidatedResources.ParsedScopes)
-            {
-                var apiScope = request.ValidatedResources.Resources.FindApiScope(parsedScope.ParsedName);
-                if (apiScope != null)
-                {
-                    var scopeVm = CreateScopeViewModel(parsedScope, apiScope, vm.ScopesConsented.Contains(parsedScope.RawValue) || model == null);
-                    apiScopes.Add(scopeVm);
-                }
-            }
-            if (ConsentOptions.EnableOfflineAccess && request.ValidatedResources.Resources.OfflineAccess)
-            {
-                apiScopes.Add(GetOfflineAccessScope(vm.ScopesConsented.Contains(IdentityServer4.IdentityServerConstants.StandardScopes.OfflineAccess) || model == null));
-            }
-            vm.ApiScopes = apiScopes;
-
-            return vm;
+            // indicate that's it ok to redirect back to authorization endpoint
+            result.RedirectUri = model.ReturnUrl;
+            result.Client = request.Client;
         }
-
-        private ScopeViewModel CreateScopeViewModel(IdentityResource identity, bool check)
+        else
         {
-            return new ScopeViewModel
-            {
-                Value = identity.Name,
-                DisplayName = identity.DisplayName ?? identity.Name,
-                Description = identity.Description,
-                Emphasize = identity.Emphasize,
-                Required = identity.Required,
-                Checked = check || identity.Required
-            };
+            // we need to redisplay the consent UI
+            result.ViewModel = await BuildViewModelAsync(model.ReturnUrl, model);
         }
 
-        public ScopeViewModel CreateScopeViewModel(ParsedScopeValue parsedScopeValue, ApiScope apiScope, bool check)
+        return result;
+    }
+
+    private async Task<ConsentViewModel> BuildViewModelAsync(string returnUrl, ConsentInputModel model = null)
+    {
+        var request = await _interaction.GetAuthorizationContextAsync(returnUrl);
+        if (request != null)
         {
-            var displayName = apiScope.DisplayName ?? apiScope.Name;
-            if (!String.IsNullOrWhiteSpace(parsedScopeValue.ParsedParameter))
+            return CreateConsentViewModel(model, returnUrl, request);
+        }
+        else
+        {
+            _logger.LogError("No consent request matching request: {0}", returnUrl);
+        }
+
+        return null;
+    }
+
+    private ConsentViewModel CreateConsentViewModel(
+        ConsentInputModel model, string returnUrl,
+        AuthorizationRequest request)
+    {
+        var vm = new ConsentViewModel
+        {
+            RememberConsent = model?.RememberConsent ?? true,
+            ScopesConsented = model?.ScopesConsented ?? Enumerable.Empty<string>(),
+            Description = model?.Description,
+
+            ReturnUrl = returnUrl,
+
+            ClientName = request.Client.ClientName ?? request.Client.ClientId,
+            ClientUrl = request.Client.ClientUri,
+            ClientLogoUrl = request.Client.LogoUri,
+            AllowRememberConsent = request.Client.AllowRememberConsent
+        };
+
+        vm.IdentityScopes = request.ValidatedResources.Resources.IdentityResources.Select(x => CreateScopeViewModel(x, vm.ScopesConsented.Contains(x.Name) || model == null)).ToArray();
+
+        var apiScopes = new List<ScopeViewModel>();
+        foreach (var parsedScope in request.ValidatedResources.ParsedScopes)
+        {
+            var apiScope = request.ValidatedResources.Resources.FindApiScope(parsedScope.ParsedName);
+            if (apiScope != null)
             {
-                displayName += ":" + parsedScopeValue.ParsedParameter;
+                var scopeVm = CreateScopeViewModel(parsedScope, apiScope, vm.ScopesConsented.Contains(parsedScope.RawValue) || model == null);
+                apiScopes.Add(scopeVm);
             }
-
-            return new ScopeViewModel
-            {
-                Value = parsedScopeValue.RawValue,
-                DisplayName = displayName,
-                Description = apiScope.Description,
-                Emphasize = apiScope.Emphasize,
-                Required = apiScope.Required,
-                Checked = check || apiScope.Required
-            };
         }
-
-        private ScopeViewModel GetOfflineAccessScope(bool check)
+        if (ConsentOptions.EnableOfflineAccess && request.ValidatedResources.Resources.OfflineAccess)
         {
-            return new ScopeViewModel
-            {
-                Value = IdentityServer4.IdentityServerConstants.StandardScopes.OfflineAccess,
-                DisplayName = _sharedLocalizer["Offline Access"],
-                Description = _sharedLocalizer["Access to your applications and resources, even when you are offline"],
-                Emphasize = true,
-                Checked = check
-            };
+            apiScopes.Add(GetOfflineAccessScope(vm.ScopesConsented.Contains(IdentityServer4.IdentityServerConstants.StandardScopes.OfflineAccess) || model == null));
         }
+        vm.ApiScopes = apiScopes;
+
+        return vm;
+    }
+
+    private ScopeViewModel CreateScopeViewModel(IdentityResource identity, bool check)
+    {
+        return new ScopeViewModel
+        {
+            Value = identity.Name,
+            DisplayName = identity.DisplayName ?? identity.Name,
+            Description = identity.Description,
+            Emphasize = identity.Emphasize,
+            Required = identity.Required,
+            Checked = check || identity.Required
+        };
+    }
+
+    public ScopeViewModel CreateScopeViewModel(ParsedScopeValue parsedScopeValue, ApiScope apiScope, bool check)
+    {
+        var displayName = apiScope.DisplayName ?? apiScope.Name;
+        if (!String.IsNullOrWhiteSpace(parsedScopeValue.ParsedParameter))
+        {
+            displayName += ":" + parsedScopeValue.ParsedParameter;
+        }
+
+        return new ScopeViewModel
+        {
+            Value = parsedScopeValue.RawValue,
+            DisplayName = displayName,
+            Description = apiScope.Description,
+            Emphasize = apiScope.Emphasize,
+            Required = apiScope.Required,
+            Checked = check || apiScope.Required
+        };
+    }
+
+    private ScopeViewModel GetOfflineAccessScope(bool check)
+    {
+        return new ScopeViewModel
+        {
+            Value = IdentityServer4.IdentityServerConstants.StandardScopes.OfflineAccess,
+            DisplayName = _sharedLocalizer["Offline Access"],
+            Description = _sharedLocalizer["Access to your applications and resources, even when you are offline"],
+            Emphasize = true,
+            Checked = check
+        };
     }
 }
