@@ -1,8 +1,4 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Reflection;
-using System.Threading.Tasks;
 using IdentityServer4.Configuration;
 using IdentityServer4.Events;
 using IdentityServer4.Extensions;
@@ -12,228 +8,226 @@ using IdentityServer4.Validation;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Localization;
-using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using StsServerIdentity.Filters;
 using StsServerIdentity.Models;
 using StsServerIdentity.Resources;
 
-namespace StsServerIdentity.Controllers
+namespace StsServerIdentity.Controllers;
+
+[Authorize]
+[SecurityHeaders]
+public class DeviceController : Controller
 {
-    [Authorize]
-    [SecurityHeaders]
-    public class DeviceController : Controller
+    private readonly IDeviceFlowInteractionService _interaction;
+    private readonly IEventService _events;
+    private readonly IOptions<IdentityServerOptions> _options;
+    private readonly ILogger<DeviceController> _logger;
+    private readonly IStringLocalizer _sharedLocalizer;
+
+    public DeviceController(
+        IDeviceFlowInteractionService interaction,
+        IEventService eventService,
+        IOptions<IdentityServerOptions> options,
+        ILogger<DeviceController> logger,
+        IStringLocalizerFactory factory)
     {
-        private readonly IDeviceFlowInteractionService _interaction;
-        private readonly IEventService _events;
-        private readonly IOptions<IdentityServerOptions> _options;
-        private readonly ILogger<DeviceController> _logger;
-        private readonly IStringLocalizer _sharedLocalizer;
+        _interaction = interaction;
+        _events = eventService;
+        _options = options;
+        _logger = logger;
 
-        public DeviceController(
-            IDeviceFlowInteractionService interaction,
-            IEventService eventService,
-            IOptions<IdentityServerOptions> options,
-            ILogger<DeviceController> logger,
-            IStringLocalizerFactory factory)
+        var type = typeof(SharedResource);
+        var assemblyName = new AssemblyName(type.GetTypeInfo().Assembly.FullName);
+        _sharedLocalizer = factory.Create("SharedResource", assemblyName.Name);
+    }
+
+    [HttpGet]
+    public async Task<IActionResult> Index()
+    {
+        string userCodeParamName = _options.Value.UserInteraction.DeviceVerificationUserCodeParameter;
+        string userCode = Request.Query[userCodeParamName];
+        if (string.IsNullOrWhiteSpace(userCode)) return View("UserCodeCapture");
+
+        var vm = await BuildViewModelAsync(userCode);
+        if (vm == null) return View("Error");
+
+        vm.ConfirmUserCode = true;
+        return View("UserCodeConfirmation", vm);
+    }
+
+    [HttpPost]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> UserCodeCapture(string userCode)
+    {
+        var vm = await BuildViewModelAsync(userCode);
+        if (vm == null) return View("Error");
+
+        return View("UserCodeConfirmation", vm);
+    }
+
+    [HttpPost]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> Callback(DeviceAuthorizationInputModel model)
+    {
+        if (model == null) throw new ArgumentNullException(nameof(model));
+
+        var result = await ProcessConsent(model);
+        if (result.HasValidationError) return View("Error");
+
+        return View("Success");
+    }
+
+    private async Task<ProcessConsentResult> ProcessConsent(DeviceAuthorizationInputModel model)
+    {
+        var result = new ProcessConsentResult();
+
+        var request = await _interaction.GetAuthorizationContextAsync(model.UserCode);
+        if (request == null) return result;
+
+        ConsentResponse grantedConsent = null;
+
+        // user clicked 'no' - send back the standard 'access_denied' response
+        if (model.Button == "no")
         {
-            _interaction = interaction;
-            _events = eventService;
-            _options = options;
-            _logger = logger;
+            grantedConsent = new ConsentResponse { Error = AuthorizationError.AccessDenied };
 
-            var type = typeof(SharedResource);
-            var assemblyName = new AssemblyName(type.GetTypeInfo().Assembly.FullName);
-            _sharedLocalizer = factory.Create("SharedResource", assemblyName.Name);
+            // emit event
+            await _events.RaiseAsync(new ConsentDeniedEvent(User.GetSubjectId(), request.Client.ClientId, request.ValidatedResources.RawScopeValues));
         }
-
-        [HttpGet]
-        public async Task<IActionResult> Index()
+        // user clicked 'yes' - validate the data
+        else if (model.Button == "yes")
         {
-            string userCodeParamName = _options.Value.UserInteraction.DeviceVerificationUserCodeParameter;
-            string userCode = Request.Query[userCodeParamName];
-            if (string.IsNullOrWhiteSpace(userCode)) return View("UserCodeCapture");
-
-            var vm = await BuildViewModelAsync(userCode);
-            if (vm == null) return View("Error");
-
-            vm.ConfirmUserCode = true;
-            return View("UserCodeConfirmation", vm);
-        }
-
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> UserCodeCapture(string userCode)
-        {
-            var vm = await BuildViewModelAsync(userCode);
-            if (vm == null) return View("Error");
-
-            return View("UserCodeConfirmation", vm);
-        }
-
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Callback(DeviceAuthorizationInputModel model)
-        {
-            if (model == null) throw new ArgumentNullException(nameof(model));
-
-            var result = await ProcessConsent(model);
-            if (result.HasValidationError) return View("Error");
-
-            return View("Success");
-        }
-
-        private async Task<ProcessConsentResult> ProcessConsent(DeviceAuthorizationInputModel model)
-        {
-            var result = new ProcessConsentResult();
-
-            var request = await _interaction.GetAuthorizationContextAsync(model.UserCode);
-            if (request == null) return result;
-
-            ConsentResponse grantedConsent = null;
-
-            // user clicked 'no' - send back the standard 'access_denied' response
-            if (model.Button == "no")
+            // if the user consented to some scope, build the response model
+            if (model.ScopesConsented != null && model.ScopesConsented.Any())
             {
-                grantedConsent = new ConsentResponse { Error = AuthorizationError.AccessDenied };
+                var scopes = model.ScopesConsented;
+                if (ConsentOptions.EnableOfflineAccess == false)
+                {
+                    scopes = scopes.Where(x => x != IdentityServer4.IdentityServerConstants.StandardScopes.OfflineAccess);
+                }
+
+                grantedConsent = new ConsentResponse
+                {
+                    RememberConsent = model.RememberConsent,
+                    ScopesValuesConsented = scopes.ToArray(),
+                    Description = model.Description
+                };
 
                 // emit event
-                await _events.RaiseAsync(new ConsentDeniedEvent(User.GetSubjectId(), request.Client.ClientId, request.ValidatedResources.RawScopeValues));
-            }
-            // user clicked 'yes' - validate the data
-            else if (model.Button == "yes")
-            {
-                // if the user consented to some scope, build the response model
-                if (model.ScopesConsented != null && model.ScopesConsented.Any())
-                {
-                    var scopes = model.ScopesConsented;
-                    if (ConsentOptions.EnableOfflineAccess == false)
-                    {
-                        scopes = scopes.Where(x => x != IdentityServer4.IdentityServerConstants.StandardScopes.OfflineAccess);
-                    }
-
-                    grantedConsent = new ConsentResponse
-                    {
-                        RememberConsent = model.RememberConsent,
-                        ScopesValuesConsented = scopes.ToArray(),
-                        Description = model.Description
-                    };
-
-                    // emit event
-                    await _events.RaiseAsync(new ConsentGrantedEvent(User.GetSubjectId(), request.Client.ClientId, request.ValidatedResources.RawScopeValues, grantedConsent.ScopesValuesConsented, grantedConsent.RememberConsent));
-                }
-                else
-                {
-                    result.ValidationError = _sharedLocalizer["You must pick at least one permission"];
-                }
+                await _events.RaiseAsync(new ConsentGrantedEvent(User.GetSubjectId(), request.Client.ClientId, request.ValidatedResources.RawScopeValues, grantedConsent.ScopesValuesConsented, grantedConsent.RememberConsent));
             }
             else
             {
-                result.ValidationError = _sharedLocalizer["Invalid selection"];
+                result.ValidationError = _sharedLocalizer["You must pick at least one permission"];
             }
-
-            if (grantedConsent != null)
-            {
-                // communicate outcome of consent back to identityserver
-                await _interaction.HandleRequestAsync(model.UserCode, grantedConsent);
-
-                // indicate that's it ok to redirect back to authorization endpoint
-                result.RedirectUri = model.ReturnUrl;
-                result.Client = request.Client;
-            }
-            else
-            {
-                // we need to redisplay the consent UI
-                result.ViewModel = await BuildViewModelAsync(model.UserCode, model);
-            }
-
-            return result;
         }
-
-        private async Task<DeviceAuthorizationViewModel> BuildViewModelAsync(string userCode, DeviceAuthorizationInputModel model = null)
+        else
         {
-            var request = await _interaction.GetAuthorizationContextAsync(userCode);
-            if (request != null)
+            result.ValidationError = _sharedLocalizer["Invalid selection"];
+        }
+
+        if (grantedConsent != null)
+        {
+            // communicate outcome of consent back to identityserver
+            await _interaction.HandleRequestAsync(model.UserCode, grantedConsent);
+
+            // indicate that's it ok to redirect back to authorization endpoint
+            result.RedirectUri = model.ReturnUrl;
+            result.Client = request.Client;
+        }
+        else
+        {
+            // we need to redisplay the consent UI
+            result.ViewModel = await BuildViewModelAsync(model.UserCode, model);
+        }
+
+        return result;
+    }
+
+    private async Task<DeviceAuthorizationViewModel> BuildViewModelAsync(string userCode, DeviceAuthorizationInputModel model = null)
+    {
+        var request = await _interaction.GetAuthorizationContextAsync(userCode);
+        if (request != null)
+        {
+            return CreateConsentViewModel(userCode, model, request);
+        }
+
+        return null;
+    }
+
+    private DeviceAuthorizationViewModel CreateConsentViewModel(string userCode, DeviceAuthorizationInputModel model, DeviceFlowAuthorizationRequest request)
+    {
+        var vm = new DeviceAuthorizationViewModel
+        {
+            UserCode = userCode,
+            Description = model?.Description,
+
+            RememberConsent = model?.RememberConsent ?? true,
+            ScopesConsented = model?.ScopesConsented ?? Enumerable.Empty<string>(),
+
+            ClientName = request.Client.ClientName ?? request.Client.ClientId,
+            ClientUrl = request.Client.ClientUri,
+            ClientLogoUrl = request.Client.LogoUri,
+            AllowRememberConsent = request.Client.AllowRememberConsent
+        };
+
+        vm.IdentityScopes = request.ValidatedResources.Resources.IdentityResources.Select(x => CreateScopeViewModel(x, vm.ScopesConsented.Contains(x.Name) || model == null)).ToArray();
+
+        var apiScopes = new List<ScopeViewModel>();
+        foreach (var parsedScope in request.ValidatedResources.ParsedScopes)
+        {
+            var apiScope = request.ValidatedResources.Resources.FindApiScope(parsedScope.ParsedName);
+            if (apiScope != null)
             {
-                return CreateConsentViewModel(userCode, model, request);
+                var scopeVm = CreateScopeViewModel(parsedScope, apiScope, vm.ScopesConsented.Contains(parsedScope.RawValue) || model == null);
+                apiScopes.Add(scopeVm);
             }
-
-            return null;
         }
-
-        private DeviceAuthorizationViewModel CreateConsentViewModel(string userCode, DeviceAuthorizationInputModel model, DeviceFlowAuthorizationRequest request)
+        if (ConsentOptions.EnableOfflineAccess && request.ValidatedResources.Resources.OfflineAccess)
         {
-            var vm = new DeviceAuthorizationViewModel
-            {
-                UserCode = userCode,
-                Description = model?.Description,
-
-                RememberConsent = model?.RememberConsent ?? true,
-                ScopesConsented = model?.ScopesConsented ?? Enumerable.Empty<string>(),
-
-                ClientName = request.Client.ClientName ?? request.Client.ClientId,
-                ClientUrl = request.Client.ClientUri,
-                ClientLogoUrl = request.Client.LogoUri,
-                AllowRememberConsent = request.Client.AllowRememberConsent
-            };
-
-            vm.IdentityScopes = request.ValidatedResources.Resources.IdentityResources.Select(x => CreateScopeViewModel(x, vm.ScopesConsented.Contains(x.Name) || model == null)).ToArray();
-
-            var apiScopes = new List<ScopeViewModel>();
-            foreach (var parsedScope in request.ValidatedResources.ParsedScopes)
-            {
-                var apiScope = request.ValidatedResources.Resources.FindApiScope(parsedScope.ParsedName);
-                if (apiScope != null)
-                {
-                    var scopeVm = CreateScopeViewModel(parsedScope, apiScope, vm.ScopesConsented.Contains(parsedScope.RawValue) || model == null);
-                    apiScopes.Add(scopeVm);
-                }
-            }
-            if (ConsentOptions.EnableOfflineAccess && request.ValidatedResources.Resources.OfflineAccess)
-            {
-                apiScopes.Add(GetOfflineAccessScope(vm.ScopesConsented.Contains(IdentityServer4.IdentityServerConstants.StandardScopes.OfflineAccess) || model == null));
-            }
-            vm.ApiScopes = apiScopes;
-
-            return vm;
+            apiScopes.Add(GetOfflineAccessScope(vm.ScopesConsented.Contains(IdentityServer4.IdentityServerConstants.StandardScopes.OfflineAccess) || model == null));
         }
+        vm.ApiScopes = apiScopes;
 
-        private ScopeViewModel CreateScopeViewModel(IdentityResource identity, bool check)
+        return vm;
+    }
+
+    private ScopeViewModel CreateScopeViewModel(IdentityResource identity, bool check)
+    {
+        return new ScopeViewModel
         {
-            return new ScopeViewModel
-            {
-                Value = identity.Name,
-                DisplayName = identity.DisplayName ?? identity.Name,
-                Description = identity.Description,
-                Emphasize = identity.Emphasize,
-                Required = identity.Required,
-                Checked = check || identity.Required
-            };
-        }
+            Value = identity.Name,
+            DisplayName = identity.DisplayName ?? identity.Name,
+            Description = identity.Description,
+            Emphasize = identity.Emphasize,
+            Required = identity.Required,
+            Checked = check || identity.Required
+        };
+    }
 
-        public ScopeViewModel CreateScopeViewModel(ParsedScopeValue parsedScopeValue, ApiScope apiScope, bool check)
+    public ScopeViewModel CreateScopeViewModel(ParsedScopeValue parsedScopeValue, ApiScope apiScope, bool check)
+    {
+        return new ScopeViewModel
         {
-            return new ScopeViewModel
-            {
-                Value = parsedScopeValue.RawValue,
-                // todo: use the parsed scope value in the display?
-                DisplayName = apiScope.DisplayName ?? apiScope.Name,
-                Description = apiScope.Description,
-                Emphasize = apiScope.Emphasize,
-                Required = apiScope.Required,
-                Checked = check || apiScope.Required
-            };
-        }
-        private ScopeViewModel GetOfflineAccessScope(bool check)
+            Value = parsedScopeValue.RawValue,
+            // todo: use the parsed scope value in the display?
+            DisplayName = apiScope.DisplayName ?? apiScope.Name,
+            Description = apiScope.Description,
+            Emphasize = apiScope.Emphasize,
+            Required = apiScope.Required,
+            Checked = check || apiScope.Required
+        };
+    }
+    private ScopeViewModel GetOfflineAccessScope(bool check)
+    {
+        return new ScopeViewModel
         {
-            return new ScopeViewModel
-            {
-                Value = IdentityServer4.IdentityServerConstants.StandardScopes.OfflineAccess,
-                DisplayName = _sharedLocalizer["Offline Access"],
-                Description = _sharedLocalizer["Access to your applications and resources, even when you are offline"],
-                Emphasize = true,
-                Checked = check
-            };
-        }
+            Value = IdentityServer4.IdentityServerConstants.StandardScopes.OfflineAccess,
+            DisplayName = _sharedLocalizer["Offline Access"],
+            Description = _sharedLocalizer["Access to your applications and resources, even when you are offline"],
+            Emphasize = true,
+            Checked = check
+        };
     }
 }
