@@ -3,26 +3,25 @@ using IdentityServer4.Models;
 using Microsoft.AspNetCore.Mvc;
 using StsServerIdentity.Models.AccountViewModels;
 
-namespace StsServerIdentity
+namespace StsServerIdentity;
+
+public static class Extensions
 {
-    public static class Extensions
+    /// <summary>
+    /// Checks if the redirect URI is for a native client.
+    /// </summary>
+    /// <returns></returns>
+    public static bool IsNativeClient(this AuthorizationRequest context)
     {
-        /// <summary>
-        /// Checks if the redirect URI is for a native client.
-        /// </summary>
-        /// <returns></returns>
-        public static bool IsNativeClient(this AuthorizationRequest context)
-        {
-            return !context.RedirectUri.StartsWith("https", StringComparison.Ordinal)
-               && !context.RedirectUri.StartsWith("http", StringComparison.Ordinal);
-        }
+        return !context.RedirectUri.StartsWith("https", StringComparison.Ordinal)
+           && !context.RedirectUri.StartsWith("http", StringComparison.Ordinal);
+    }
 
-        public static IActionResult LoadingPage(this Controller controller, string viewName, string redirectUri)
-        {
-            controller.HttpContext.Response.StatusCode = 200;
-            controller.HttpContext.Response.Headers["Location"] = "";
+    public static IActionResult LoadingPage(this Controller controller, string viewName, string redirectUri)
+    {
+        controller.HttpContext.Response.StatusCode = 200;
+        controller.HttpContext.Response.Headers["Location"] = "";
 
-            return controller.View(viewName, new RedirectViewModel { RedirectUrl = redirectUri });
-        }
+        return controller.View(viewName, new RedirectViewModel { RedirectUrl = redirectUri });
     }
 }
