@@ -4,42 +4,41 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
 
-namespace StsServerIdentity.Filters
+namespace StsServerIdentity.Filters;
+
+public class SecurityHeadersAttribute : ActionFilterAttribute
 {
-    public class SecurityHeadersAttribute : ActionFilterAttribute
+    public override void OnResultExecuting(ResultExecutingContext context)
     {
-        public override void OnResultExecuting(ResultExecutingContext context)
+        var result = context.Result;
+        if (result is ViewResult)
         {
-            var result = context.Result;
-            if (result is ViewResult)
+            var featurePolicy = "accelerometer 'none'; camera 'none'; geolocation 'none'; gyroscope 'none'; magnetometer 'none'; microphone 'none'; payment 'none'; usb 'none'";
+
+            if (!context.HttpContext.Response.Headers.ContainsKey("feature-policy"))
             {
-                var featurePolicy = "accelerometer 'none'; camera 'none'; geolocation 'none'; gyroscope 'none'; magnetometer 'none'; microphone 'none'; payment 'none'; usb 'none'";
+                context.HttpContext.Response.Headers.Add("feature-policy", featurePolicy);
+            }
 
-                if (!context.HttpContext.Response.Headers.ContainsKey("feature-policy"))
-                {
-                    context.HttpContext.Response.Headers.Add("feature-policy", featurePolicy);
-                }
+            if (!context.HttpContext.Response.Headers.ContainsKey("X-Content-Type-Options"))
+            {
+                context.HttpContext.Response.Headers.Add("X-Content-Type-Options", "nosniff");
+            }
+            if (!context.HttpContext.Response.Headers.ContainsKey("X-Frame-Options"))
+            {
+                context.HttpContext.Response.Headers.Add("X-Frame-Options", "SAMEORIGIN");
+            }
 
-                if (!context.HttpContext.Response.Headers.ContainsKey("X-Content-Type-Options"))
-                {
-                    context.HttpContext.Response.Headers.Add("X-Content-Type-Options", "nosniff");
-                }
-                if (!context.HttpContext.Response.Headers.ContainsKey("X-Frame-Options"))
-                {
-                    context.HttpContext.Response.Headers.Add("X-Frame-Options", "SAMEORIGIN");
-                }
-
-                var csp = "script-src 'self' 'unsafe-inline';style-src 'self' 'unsafe-inline';img-src 'self' data:;font-src 'self';form-action 'self';frame-ancestors 'self';block-all-mixed-content";
-                // once for standards compliant browsers
-                if (!context.HttpContext.Response.Headers.ContainsKey("Content-Security-Policy"))
-                {
-                    context.HttpContext.Response.Headers.Add("Content-Security-Policy", csp);
-                }
-                // and once again for IE
-                if (!context.HttpContext.Response.Headers.ContainsKey("X-Content-Security-Policy"))
-                {
-                    context.HttpContext.Response.Headers.Add("X-Content-Security-Policy", csp);
-                }
+            var csp = "script-src 'self' 'unsafe-inline';style-src 'self' 'unsafe-inline';img-src 'self' data:;font-src 'self';form-action 'self';frame-ancestors 'self';block-all-mixed-content";
+            // once for standards compliant browsers
+            if (!context.HttpContext.Response.Headers.ContainsKey("Content-Security-Policy"))
+            {
+                context.HttpContext.Response.Headers.Add("Content-Security-Policy", csp);
+            }
+            // and once again for IE
+            if (!context.HttpContext.Response.Headers.ContainsKey("X-Content-Security-Policy"))
+            {
+                context.HttpContext.Response.Headers.Add("X-Content-Security-Policy", csp);
             }
         }
     }
