@@ -1,74 +1,68 @@
-﻿using System;
-using Azure.Identity;
-using Microsoft.AspNetCore;
-using Microsoft.AspNetCore.Hosting;
+﻿using Azure.Identity;
 using Microsoft.Azure.Services.AppAuthentication;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Hosting;
 using Serilog;
 using Serilog.Events;
 using Serilog.Sinks.SystemConsole.Themes;
 
-namespace StsServerIdentity
+namespace StsServerIdentity;
+
+public class Program
 {
-    public class Program
+    public static int Main(string[] args)
     {
-        public static int Main(string[] args)
+        Log.Logger = new LoggerConfiguration()
+            .MinimumLevel.Debug()
+            .MinimumLevel.Override("Microsoft", LogEventLevel.Information)
+            .Enrich.FromLogContext()
+            .WriteTo.Console()
+            .CreateLogger();
+
+        try
         {
-            Log.Logger = new LoggerConfiguration()
-                .MinimumLevel.Debug()
-                .MinimumLevel.Override("Microsoft", LogEventLevel.Information)
-                .Enrich.FromLogContext()
-                .WriteTo.Console()
-                .CreateLogger();
-
-            try
-            {
-                Log.Information("Starting web host");
-                CreateHostBuilder(args).Build().Run();
-                return 0;
-            }
-            catch (Exception ex)
-            {
-                Log.Fatal(ex, "Host terminated unexpectedly");
-                return 1;
-            }
-            finally
-            {
-                Log.CloseAndFlush();
-            }
+            Log.Information("Starting web host");
+            CreateHostBuilder(args).Build().Run();
+            return 0;
         }
-
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureAppConfiguration((context, config) =>
-                {
-                    var builder = config.Build();
-                    var keyVaultEndpoint = builder["AzureKeyVaultEndpoint"];
-                    if (!string.IsNullOrEmpty(keyVaultEndpoint))
-                    {
-                        var azureServiceTokenProvider = new AzureServiceTokenProvider();
-                        config.AddAzureKeyVault(new Uri(keyVaultEndpoint), new DefaultAzureCredential());
-                    }
-                    else
-                    {
-                        IHostEnvironment env = context.HostingEnvironment;
-
-                        config.AddJsonFile("appsettings.json", optional: false, reloadOnChange: true)
-                            .AddJsonFile($"appsettings.{env.EnvironmentName}.json", optional: true, reloadOnChange: true)
-                            .AddEnvironmentVariables();
-                        //.AddUserSecrets("your user secret....");
-                    }
-                })
-                .ConfigureWebHostDefaults(webBuilder =>
-                {
-                    webBuilder.UseStartup<Startup>()
-                        .UseSerilog((hostingContext, loggerConfiguration) => loggerConfiguration
-                        .ReadFrom.Configuration(hostingContext.Configuration)
-                        .Enrich.FromLogContext()
-                        .WriteTo.File("../StsLogs.txt")
-                        .WriteTo.Console(theme: AnsiConsoleTheme.Code)
-                );
-                });
+        catch (Exception ex)
+        {
+            Log.Fatal(ex, "Host terminated unexpectedly");
+            return 1;
+        }
+        finally
+        {
+            Log.CloseAndFlush();
+        }
     }
+
+    public static IHostBuilder CreateHostBuilder(string[] args) =>
+   Host.CreateDefaultBuilder(args)
+       .ConfigureAppConfiguration((context, config) =>
+       {
+           var builder = config.Build();
+           var keyVaultEndpoint = builder["AzureKeyVaultEndpoint"];
+           if (!string.IsNullOrEmpty(keyVaultEndpoint))
+           {
+               var azureServiceTokenProvider = new AzureServiceTokenProvider();
+               config.AddAzureKeyVault(new Uri(keyVaultEndpoint), new DefaultAzureCredential());
+           }
+           else
+           {
+               IHostEnvironment env = context.HostingEnvironment;
+
+               config.AddJsonFile("appsettings.json", optional: false, reloadOnChange: true)
+                   .AddJsonFile($"appsettings.{env.EnvironmentName}.json", optional: true, reloadOnChange: true)
+                   .AddEnvironmentVariables();
+               //.AddUserSecrets("your user secret....");
+           }
+
+       })
+       .UseSerilog((hostingContext, loggerConfiguration) => loggerConfiguration
+           .ReadFrom.Configuration(hostingContext.Configuration)
+           .Enrich.FromLogContext()
+           .WriteTo.Console(theme: AnsiConsoleTheme.Code)
+           .WriteTo.File("../stsLogs.txt"))
+       .ConfigureWebHostDefaults(webBuilder =>
+       {
+           webBuilder.UseStartup<Startup>();
+       });
 }
