@@ -1,37 +1,36 @@
 ﻿using Microsoft.Extensions.Localization;
 using System.Reflection;
 
-namespace StsServerIdentity.Resources
+namespace StsServerIdentity.Resources;
+
+public class LocService
 {
-    public class LocService
+    private readonly IStringLocalizer _localizer;
+
+    public LocService(IStringLocalizerFactory factory)
     {
-        private readonly IStringLocalizer _localizer;
+        var type = typeof(SharedResource);
+        var assemblyName = new AssemblyName(type.GetTypeInfo().Assembly.FullName);
+        _localizer = factory.Create("SharedResource", assemblyName.Name);
+    }
 
-        public LocService(IStringLocalizerFactory factory)
-        {
-            var type = typeof(SharedResource);
-            var assemblyName = new AssemblyName(type.GetTypeInfo().Assembly.FullName);
-            _localizer = factory.Create("SharedResource", assemblyName.Name);
-        }
+    public LocalizedString GetLocalizedHtmlString(string key)
+    {
+        return _localizer[key];
+    }
 
-        public LocalizedString GetLocalizedHtmlString(string key)
+    public LocalizedString GetLocalizedHtmlStringAllowNull(string key)
+    {
+        if (!string.IsNullOrWhiteSpace(key))
         {
             return _localizer[key];
         }
 
-        public LocalizedString GetLocalizedHtmlStringAllowNull(string key)
-        {
-            if (!string.IsNullOrWhiteSpace(key))
-            {
-                return _localizer[key];
-            }
+        return new LocalizedString(key, string.Empty);
+    }
 
-            return new LocalizedString(key, string.Empty);
-        }
-
-        public LocalizedString GetLocalizedHtmlString(string key, string parameter)
-        {
-            return _localizer[key, parameter];
-        }
+    public LocalizedString GetLocalizedHtmlString(string key, string parameter)
+    {
+        return _localizer[key, parameter];
     }
 }
