@@ -1,9 +1,8 @@
-﻿namespace StsServerIdentity.Resources
+﻿namespace StsServerIdentity.Resources;
+
+/// <summary>
+/// Dummy class to group shared resources
+/// </summary>
+public class SharedResource
 {
-    /// <summary>
-    /// Dummy class to group shared resources
-    /// </summary>
-    public class SharedResource
-    {
-    }
 }
