@@ -2,32 +2,30 @@
 using SendGrid;
 using SendGrid.Helpers.Mail;
 using StsServerIdentity.Models;
-using System.Threading.Tasks;
 
-namespace StsServerIdentity.Services
+namespace StsServerIdentity.Services;
+
+public class EmailSender : IEmailSender
 {
-    public class EmailSender : IEmailSender
+    private readonly IOptions<EmailSettings> _optionsEmailSettings;
+
+    public EmailSender(IOptions<EmailSettings> optionsEmailSettings)
     {
-        private readonly IOptions<EmailSettings> _optionsEmailSettings;
+        _optionsEmailSettings = optionsEmailSettings;
+    }
 
-        public EmailSender(IOptions<EmailSettings> optionsEmailSettings)
-        {
-            _optionsEmailSettings = optionsEmailSettings;
-        }
+    public async Task SendEmail(string email, string subject, string message, string toUsername)
+    {
+        var client = new SendGridClient(_optionsEmailSettings.Value.SendGridApiKey);
+        var msg = new SendGridMessage();
+        msg.SetFrom(new EmailAddress(_optionsEmailSettings.Value.SenderEmailAddress, "damienbod"));
+        msg.AddTo(new EmailAddress(email, toUsername));
+        msg.SetSubject(subject);
+        msg.AddContent(MimeType.Text, message);
+        //msg.AddContent(MimeType.Html, message);
 
-        public async Task SendEmail(string email, string subject, string message, string toUsername)
-        {
-            var client = new SendGridClient(_optionsEmailSettings.Value.SendGridApiKey);
-            var msg = new SendGridMessage();
-            msg.SetFrom(new EmailAddress(_optionsEmailSettings.Value.SenderEmailAddress, "damienbod"));
-            msg.AddTo(new EmailAddress(email, toUsername));
-            msg.SetSubject(subject);
-            msg.AddContent(MimeType.Text, message);
-            //msg.AddContent(MimeType.Html, message);
+        msg.SetReplyTo(new EmailAddress(_optionsEmailSettings.Value.SenderEmailAddress, "damienbod"));
 
-            msg.SetReplyTo(new EmailAddress(_optionsEmailSettings.Value.SenderEmailAddress, "damienbod"));
-
-            var response = await client.SendEmailAsync(msg);
-        }
+        var response = await client.SendEmailAsync(msg);
     }
 }
