@@ -1,17 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.IO;
+﻿using System.Globalization;
 using System.Reflection;
 using System.Security.Cryptography.X509Certificates;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Localization;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
 using IdentityServer4.Services;
@@ -22,313 +14,314 @@ using StsServerIdentity.Services;
 using StsServerIdentity.Filters;
 using StsServerIdentity.Services.Certificate;
 using Serilog;
-using Microsoft.AspNetCore.Http;
 using Fido2NetLib;
 using Microsoft.AspNetCore.Authorization;
-using System.Threading.Tasks;
 
-namespace StsServerIdentity
+namespace StsServerIdentity;
+
+public class Startup
 {
-    public class Startup
+    public Startup(IConfiguration configuration, IWebHostEnvironment env)
     {
-        public Startup(IConfiguration configuration, IWebHostEnvironment env)
+        Configuration = configuration;
+        _environment = env;
+    }
+
+    public IConfiguration Configuration { get; }
+    public IWebHostEnvironment _environment { get; }
+
+    public void ConfigureServices(IServiceCollection services)
+    {
+        var stsConfig = Configuration.GetSection("StsConfig");
+        var _clientId = Configuration["MicrosoftClientId"];
+        var _clientSecret = Configuration["MircosoftClientSecret"];
+
+        services.AddDbContext<ApplicationDbContext>(options =>
+            options.UseSqlite(Configuration.GetConnectionString("DefaultConnection")));
+
+        services.Configure<StsConfig>(Configuration.GetSection("StsConfig"));
+        services.Configure<EmailSettings>(Configuration.GetSection("EmailSettings"));
+        services.AddTransient<IProfileService, IdentityWithAdditionalClaimsProfileService>();
+        services.AddTransient<IEmailSender, EmailSender>();
+        services.AddSingleton<IAuthorizationHandler, IsAdminHandler>();
+
+        var x509Certificate2Certs = GetCertificates(_environment, Configuration)
+            .GetAwaiter().GetResult();
+
+        services.Configure<CookiePolicyOptions>(options =>
         {
-            Configuration = configuration;
-            _environment = env;
-        }
+            options.MinimumSameSitePolicy = SameSiteMode.Unspecified;
+            options.OnAppendCookie = cookieContext =>
+                CheckSameSite(cookieContext.Context, cookieContext.CookieOptions);
+            options.OnDeleteCookie = cookieContext =>
+                CheckSameSite(cookieContext.Context, cookieContext.CookieOptions);
+        });
 
-        public IConfiguration Configuration { get; }
-        public IWebHostEnvironment _environment { get; }
+        AddLocalizationConfigurations(services);
 
-        public void ConfigureServices(IServiceCollection services)
+        services.AddCors(options =>
         {
-            var stsConfig = Configuration.GetSection("StsConfig");
-            var _clientId = Configuration["MicrosoftClientId"];
-            var _clientSecret = Configuration["MircosoftClientSecret"];
-
-            services.AddDbContext<ApplicationDbContext>(options =>
-                options.UseSqlite(Configuration.GetConnectionString("DefaultConnection")));
-
-            services.Configure<StsConfig>(Configuration.GetSection("StsConfig"));
-            services.Configure<EmailSettings>(Configuration.GetSection("EmailSettings"));
-            services.AddTransient<IProfileService, IdentityWithAdditionalClaimsProfileService>();
-            services.AddTransient<IEmailSender, EmailSender>();
-            services.AddSingleton<IAuthorizationHandler, IsAdminHandler>();
-
-            var x509Certificate2Certs = GetCertificates(_environment, Configuration)
-                .GetAwaiter().GetResult();
-
-            services.Configure<CookiePolicyOptions>(options =>
-            {
-                options.MinimumSameSitePolicy = SameSiteMode.Unspecified;
-                options.OnAppendCookie = cookieContext =>
-                    CheckSameSite(cookieContext.Context, cookieContext.CookieOptions);
-                options.OnDeleteCookie = cookieContext =>
-                    CheckSameSite(cookieContext.Context, cookieContext.CookieOptions);
-            });
-
-            AddLocalizationConfigurations(services);
-
-            services.AddCors(options =>
-            {
-                options.AddPolicy("AllowAllOrigins",
-                    builder =>
-                    {
-                        builder
-                            .AllowCredentials()
-                            .WithOrigins("https://localhost:44311", "https://localhost:44390", "https://localhost:44395", "https://localhost:44318")
-                            .SetIsOriginAllowedToAllowWildcardSubdomains()
-                            .AllowAnyHeader()
-                            .AllowAnyMethod();
-                    });
-            });
-
-            services.AddIdentity<ApplicationUser, IdentityRole>()
-                .AddEntityFrameworkStores<ApplicationDbContext>()
-                .AddErrorDescriber<StsIdentityErrorDescriber>()
-                .AddDefaultTokenProviders()
-                .AddTokenProvider<Fifo2UserTwoFactorTokenProvider>("FIDO2");
-
-
-            services.AddScoped<IUserClaimsPrincipalFactory<ApplicationUser>,
-                AdditionalUserClaimsPrincipalFactory>();
-
-            services.AddAuthentication()
-                 .AddOpenIdConnect("aad", "Login with Azure AD", options =>
-                 {
-                     options.Authority = $"https://login.microsoftonline.com/common";
-                     options.TokenValidationParameters = new TokenValidationParameters { ValidateIssuer = false };
-                     options.ClientId = "99eb0b9d-ca40-476e-b5ac-6f4c32bfb530";
-                     options.CallbackPath = "/signin-oidc";
-                 });
-
-            services.AddAuthorization(options =>
-            {
-                options.AddPolicy("IsAdmin", policyIsAdminRequirement =>
+            options.AddPolicy("AllowAllOrigins",
+                builder =>
                 {
-                    policyIsAdminRequirement.Requirements.Add(new IsAdminRequirement());
+                    builder
+                        .AllowCredentials()
+                        .WithOrigins("https://localhost:44311", "https://localhost:44390", "https://localhost:44395", "https://localhost:44318")
+                        .SetIsOriginAllowedToAllowWildcardSubdomains()
+                        .AllowAnyHeader()
+                        .AllowAnyMethod();
                 });
-            });
-            services.AddControllersWithViews(options =>
-                {
-                    options.Filters.Add(new SecurityHeadersAttribute());
-                })
-                .AddViewLocalization()
-                .AddDataAnnotationsLocalization(options =>
-                {
-                    options.DataAnnotationLocalizerProvider = (type, factory) =>
-                    {
-                        var assemblyName = new AssemblyName(typeof(SharedResource).GetTypeInfo().Assembly.FullName);
-                        return factory.Create("SharedResource", assemblyName.Name);
-                    };
-                })
-                .AddNewtonsoftJson();
+        });
 
-            services.AddIdentityServer()
-                .AddSigningCredential(x509Certificate2Certs.ActiveCertificate)
-                .AddInMemoryIdentityResources(Config.GetIdentityResources())
-                .AddInMemoryApiResources(Config.GetApiResources())
-                // .leg
-                .AddInMemoryApiScopes(Config.GetApiScopes())
-                .AddInMemoryClients(Config.GetClients())
-                .AddAspNetIdentity<ApplicationUser>()
-                .AddProfileService<IdentityWithAdditionalClaimsProfileService>();
+        services.AddIdentity<ApplicationUser, IdentityRole>()
+            .AddEntityFrameworkStores<ApplicationDbContext>()
+            .AddErrorDescriber<StsIdentityErrorDescriber>()
+            .AddDefaultTokenProviders()
+            .AddTokenProvider<Fido2UserTwoFactorTokenProvider>("FIDO2");
 
-            services.Configure<Fido2Configuration>(Configuration.GetSection("fido2"));
-            services.Configure<Fido2MdsConfiguration>(Configuration.GetSection("fido2mds"));
-            services.AddScoped<Fido2Storage>();
-            // Adds a default in-memory implementation of IDistributedCache.
-            services.AddDistributedMemoryCache();
-            services.AddSession(options =>
-            {
-                options.IdleTimeout = TimeSpan.FromMinutes(2);
-                options.Cookie.HttpOnly = true;
-                options.Cookie.SameSite = SameSiteMode.None;
-                options.Cookie.SecurePolicy = CookieSecurePolicy.Always;
-            });
-        }
+        //builder.Services.AddDefaultIdentity<IdentityUser>(options => options.SignIn.RequireConfirmedAccount = false)
+        //    .AddEntityFrameworkStores<ApplicationDbContext>()
+        //    .AddTokenProvider<Fido2UserTwoFactorTokenProvider>("FIDO2");
 
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+
+        services.AddScoped<IUserClaimsPrincipalFactory<ApplicationUser>,
+            AdditionalUserClaimsPrincipalFactory>();
+
+        services.AddAuthentication()
+             .AddOpenIdConnect("aad", "Login with Azure AD", options =>
+             {
+                 options.Authority = $"https://login.microsoftonline.com/common";
+                 options.TokenValidationParameters = new TokenValidationParameters { ValidateIssuer = false };
+                 options.ClientId = "99eb0b9d-ca40-476e-b5ac-6f4c32bfb530";
+                 options.CallbackPath = "/signin-oidc";
+             });
+
+        services.AddAuthorization(options =>
         {
-            if (env.IsDevelopment())
+            options.AddPolicy("IsAdmin", policyIsAdminRequirement =>
             {
-                app.UseDeveloperExceptionPage();
-            }
-            else
+                policyIsAdminRequirement.Requirements.Add(new IsAdminRequirement());
+            });
+        });
+        services.AddControllersWithViews(options =>
             {
-                app.UseExceptionHandler("/Error");
-                app.UseHsts(hsts => hsts.MaxAge(365).IncludeSubdomains());
-            }
+                options.Filters.Add(new SecurityHeadersAttribute());
+            })
+            .AddViewLocalization()
+            .AddDataAnnotationsLocalization(options =>
+            {
+                options.DataAnnotationLocalizerProvider = (type, factory) =>
+                {
+                    var assemblyName = new AssemblyName(typeof(SharedResource).GetTypeInfo().Assembly.FullName);
+                    return factory.Create("SharedResource", assemblyName.Name);
+                };
+            })
+            .AddNewtonsoftJson();
 
-            app.UseCors("AllowAllOrigins");
+        services.AddIdentityServer()
+            .AddSigningCredential(x509Certificate2Certs.ActiveCertificate)
+            .AddInMemoryIdentityResources(Config.GetIdentityResources())
+            .AddInMemoryApiResources(Config.GetApiResources())
+            // .leg
+            .AddInMemoryApiScopes(Config.GetApiScopes())
+            .AddInMemoryClients(Config.GetClients())
+            .AddAspNetIdentity<ApplicationUser>()
+            .AddProfileService<IdentityWithAdditionalClaimsProfileService>();
 
+        services.Configure<Fido2Configuration>(Configuration.GetSection("fido2"));
+        services.AddScoped<Fido2Store>();
+
+        // Adds a default in-memory implementation of IDistributedCache.
+        services.AddDistributedMemoryCache();
+        services.AddSession(options =>
+        {
+            options.IdleTimeout = TimeSpan.FromMinutes(2);
+            options.Cookie.HttpOnly = true;
+            options.Cookie.SameSite = SameSiteMode.None;
+            options.Cookie.SecurePolicy = CookieSecurePolicy.Always;
+        });
+    }
+
+    public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+    {
+        if (env.IsDevelopment())
+        {
+            app.UseDeveloperExceptionPage();
+        }
+        else
+        {
+            app.UseExceptionHandler("/Error");
             app.UseHsts(hsts => hsts.MaxAge(365).IncludeSubdomains());
-            //app.UseXContentTypeOptions();
-            //app.UseReferrerPolicy(opts => opts.NoReferrer());
-            //app.UseXXssProtection(options => options.EnabledWithBlockMode());
+        }
 
-            var stsConfig = Configuration.GetSection("StsConfig");
-            var angularClient2Url = stsConfig["AngularClient2Url"];
-            var angularClientUrl = stsConfig["AngularClientUrl"];
+        app.UseCors("AllowAllOrigins");
 
-            app.UseCsp(opts => opts
-                .BlockAllMixedContent()
-                .StyleSources(s => s.Self())
-                .StyleSources(s => s.UnsafeInline())
-                .FontSources(s => s.Self())
-                .FrameAncestors(s => s.Self())
-                .FrameAncestors(s => s.CustomSources(
-                   "https://localhost:44311", "https://localhost:44390", "https://localhost:44395", "https://localhost:44318")
-                 )
-                .ImageSources(imageSrc => imageSrc.Self())
-                .ImageSources(imageSrc => imageSrc.CustomSources("data:"))
-                .ScriptSources(s => s.Self())
-                .ScriptSources(s => s.UnsafeInline())
-            );
+        app.UseHsts(hsts => hsts.MaxAge(365).IncludeSubdomains());
+        //app.UseXContentTypeOptions();
+        //app.UseReferrerPolicy(opts => opts.NoReferrer());
+        //app.UseXXssProtection(options => options.EnabledWithBlockMode());
 
-            var locOptions = app.ApplicationServices.GetService<IOptions<RequestLocalizationOptions>>();
-            app.UseRequestLocalization(locOptions.Value);
+        var stsConfig = Configuration.GetSection("StsConfig");
+        var angularClient2Url = stsConfig["AngularClient2Url"];
+        var angularClientUrl = stsConfig["AngularClientUrl"];
 
-            // https://nblumhardt.com/2019/10/serilog-in-aspnetcore-3/
-            // https://nblumhardt.com/2019/10/serilog-mvc-logging/
-            app.UseSerilogRequestLogging();
+        app.UseCsp(opts => opts
+            .BlockAllMixedContent()
+            .StyleSources(s => s.Self())
+            .StyleSources(s => s.UnsafeInline())
+            .FontSources(s => s.Self())
+            .FrameAncestors(s => s.Self())
+            .FrameAncestors(s => s.CustomSources(
+               "https://localhost:44311", "https://localhost:44390", "https://localhost:44395", "https://localhost:44318")
+             )
+            .ImageSources(imageSrc => imageSrc.Self())
+            .ImageSources(imageSrc => imageSrc.CustomSources("data:"))
+            .ScriptSources(s => s.Self())
+            .ScriptSources(s => s.UnsafeInline())
+        );
 
-            app.UseStaticFiles(new StaticFileOptions()
+        var locOptions = app.ApplicationServices.GetService<IOptions<RequestLocalizationOptions>>();
+        app.UseRequestLocalization(locOptions.Value);
+
+        // https://nblumhardt.com/2019/10/serilog-in-aspnetcore-3/
+        // https://nblumhardt.com/2019/10/serilog-mvc-logging/
+        app.UseSerilogRequestLogging();
+
+        app.UseStaticFiles(new StaticFileOptions()
+        {
+            OnPrepareResponse = context =>
             {
-                OnPrepareResponse = context =>
+                if (context.Context.Response.Headers["feature-policy"].Count == 0)
                 {
-                    if (context.Context.Response.Headers["feature-policy"].Count == 0)
-                    {
-                        var featurePolicy = "accelerometer 'none'; camera 'none'; geolocation 'none'; gyroscope 'none'; magnetometer 'none'; microphone 'none'; payment 'none'; usb 'none'";
+                    var featurePolicy = "accelerometer 'none'; camera 'none'; geolocation 'none'; gyroscope 'none'; magnetometer 'none'; microphone 'none'; payment 'none'; usb 'none'";
 
-                        context.Context.Response.Headers["feature-policy"] = featurePolicy;
-                    }
-
-                    if (context.Context.Response.Headers["X-Content-Security-Policy"].Count == 0)
-                    {
-                        var csp = "script-src 'self';style-src 'self';img-src 'self' data:;font-src 'self';form-action 'self';frame-ancestors 'self';block-all-mixed-content";
-                        // IE
-                        context.Context.Response.Headers["X-Content-Security-Policy"] = csp;
-                    }
+                    context.Context.Response.Headers["feature-policy"] = featurePolicy;
                 }
-            });
 
-            app.UseRouting();
-
-            app.UseIdentityServer();
-            app.UseAuthorization();
-
-            app.UseSession();
-
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapControllerRoute(
-                    name: "default",
-                    pattern: "{controller=Home}/{action=Index}/{id?}");
-            });
-        }
-
-        private static async Task<(X509Certificate2 ActiveCertificate, X509Certificate2 SecondaryCertificate)> GetCertificates(IWebHostEnvironment environment, IConfiguration configuration)
-        {
-            var certificateConfiguration = new CertificateConfiguration
-            {
-                // Use an Azure key vault
-                CertificateNameKeyVault = configuration["CertificateNameKeyVault"], //"StsCert",
-                KeyVaultEndpoint = configuration["AzureKeyVaultEndpoint"], // "https://damienbod.vault.azure.net"
-
-                // Use a local store with thumbprint
-                //UseLocalCertStore = Convert.ToBoolean(configuration["UseLocalCertStore"]),
-                //CertificateThumbprint = configuration["CertificateThumbprint"],
-
-                // development certificate
-                DevelopmentCertificatePfx = Path.Combine(environment.ContentRootPath, "sts_dev_cert.pfx"),
-                DevelopmentCertificatePassword = "1234" //configuration["DevelopmentCertificatePassword"] //"1234",
-            };
-
-            (X509Certificate2 ActiveCertificate, X509Certificate2 SecondaryCertificate) certs = await CertificateService.GetCertificates(
-                certificateConfiguration).ConfigureAwait(false);
-
-            return certs;
-        }
-
-        private static void AddLocalizationConfigurations(IServiceCollection services)
-        {
-            services.AddSingleton<LocService>();
-            services.AddLocalization(options => options.ResourcesPath = "Resources");
-
-            services.Configure<RequestLocalizationOptions>(
-                options =>
+                if (context.Context.Response.Headers["X-Content-Security-Policy"].Count == 0)
                 {
-                    var supportedCultures = new List<CultureInfo>
-                        {
-                            new CultureInfo("en-US"),
-                            new CultureInfo("de-DE"),
-                            new CultureInfo("de-CH"),
-                            new CultureInfo("it-IT"),
-                            new CultureInfo("gsw-CH"),
-                            new CultureInfo("fr-FR"),
-                            new CultureInfo("zh-Hans")
-                        };
+                    var csp = "script-src 'self';style-src 'self';img-src 'self' data:;font-src 'self';form-action 'self';frame-ancestors 'self';block-all-mixed-content";
+                    // IE
+                    context.Context.Response.Headers["X-Content-Security-Policy"] = csp;
+                }
+            }
+        });
 
-                    options.DefaultRequestCulture = new RequestCulture(culture: "de-DE", uiCulture: "de-DE");
-                    options.SupportedCultures = supportedCultures;
-                    options.SupportedUICultures = supportedCultures;
+        app.UseRouting();
 
-                    var providerQuery = new LocalizationQueryProvider
+        app.UseIdentityServer();
+        app.UseAuthorization();
+
+        app.UseSession();
+
+        app.UseEndpoints(endpoints =>
+        {
+            endpoints.MapControllerRoute(
+                name: "default",
+                pattern: "{controller=Home}/{action=Index}/{id?}");
+        });
+    }
+
+    private static async Task<(X509Certificate2 ActiveCertificate, X509Certificate2 SecondaryCertificate)> GetCertificates(IWebHostEnvironment environment, IConfiguration configuration)
+    {
+        var certificateConfiguration = new CertificateConfiguration
+        {
+            // Use an Azure key vault
+            CertificateNameKeyVault = configuration["CertificateNameKeyVault"], //"StsCert",
+            KeyVaultEndpoint = configuration["AzureKeyVaultEndpoint"], // "https://damienbod.vault.azure.net"
+
+            // Use a local store with thumbprint
+            //UseLocalCertStore = Convert.ToBoolean(configuration["UseLocalCertStore"]),
+            //CertificateThumbprint = configuration["CertificateThumbprint"],
+
+            // development certificate
+            DevelopmentCertificatePfx = Path.Combine(environment.ContentRootPath, "sts_dev_cert.pfx"),
+            DevelopmentCertificatePassword = "1234" //configuration["DevelopmentCertificatePassword"] //"1234",
+        };
+
+        (X509Certificate2 ActiveCertificate, X509Certificate2 SecondaryCertificate) certs = await CertificateService.GetCertificates(
+            certificateConfiguration).ConfigureAwait(false);
+
+        return certs;
+    }
+
+    private static void AddLocalizationConfigurations(IServiceCollection services)
+    {
+        services.AddSingleton<LocService>();
+        services.AddLocalization(options => options.ResourcesPath = "Resources");
+
+        services.Configure<RequestLocalizationOptions>(
+            options =>
+            {
+                var supportedCultures = new List<CultureInfo>
                     {
-                        QueryParameterName = "ui_locales"
+                        new CultureInfo("en-US"),
+                        new CultureInfo("de-DE"),
+                        new CultureInfo("de-CH"),
+                        new CultureInfo("it-IT"),
+                        new CultureInfo("gsw-CH"),
+                        new CultureInfo("fr-FR"),
+                        new CultureInfo("zh-Hans")
                     };
 
-                    options.RequestCultureProviders.Insert(0, providerQuery);
-                });
-        }
+                options.DefaultRequestCulture = new RequestCulture(culture: "de-DE", uiCulture: "de-DE");
+                options.SupportedCultures = supportedCultures;
+                options.SupportedUICultures = supportedCultures;
 
-        private static void CheckSameSite(HttpContext httpContext, CookieOptions options)
-        {
-            if (options.SameSite == SameSiteMode.None)
-            {
-                var userAgent = httpContext.Request.Headers["User-Agent"].ToString();
-                if (DisallowsSameSiteNone(userAgent))
+                var providerQuery = new LocalizationQueryProvider
                 {
-                    // For .NET Core < 3.1 set SameSite = (SameSiteMode)(-1)
-                    options.SameSite = SameSiteMode.Unspecified;
-                }
-            }
-        }
+                    QueryParameterName = "ui_locales"
+                };
 
-        private static bool DisallowsSameSiteNone(string userAgent)
+                options.RequestCultureProviders.Insert(0, providerQuery);
+            });
+    }
+
+    private static void CheckSameSite(HttpContext httpContext, CookieOptions options)
+    {
+        if (options.SameSite == SameSiteMode.None)
         {
-            // Cover all iOS based browsers here. This includes:
-            // - Safari on iOS 12 for iPhone, iPod Touch, iPad
-            // - WkWebview on iOS 12 for iPhone, iPod Touch, iPad
-            // - Chrome on iOS 12 for iPhone, iPod Touch, iPad
-            // All of which are broken by SameSite=None, because they use the iOS networking stack
-            if (userAgent.Contains("CPU iPhone OS 12") || userAgent.Contains("iPad; CPU OS 12"))
+            var userAgent = httpContext.Request.Headers["User-Agent"].ToString();
+            if (DisallowsSameSiteNone(userAgent))
             {
-                return true;
+                // For .NET Core < 3.1 set SameSite = (SameSiteMode)(-1)
+                options.SameSite = SameSiteMode.Unspecified;
             }
-
-            // Cover Mac OS X based browsers that use the Mac OS networking stack. This includes:
-            // - Safari on Mac OS X.
-            // This does not include:
-            // - Chrome on Mac OS X
-            // Because they do not use the Mac OS networking stack.
-            if (userAgent.Contains("Macintosh; Intel Mac OS X 10_14") &&
-                userAgent.Contains("Version/") && userAgent.Contains("Safari"))
-            {
-                return true;
-            }
-
-            // Cover Chrome 50-69, because some versions are broken by SameSite=None, 
-            // and none in this range require it.
-            // Note: this covers some pre-Chromium Edge versions, 
-            // but pre-Chromium Edge does not require SameSite=None.
-            if (userAgent.Contains("Chrome/5") || userAgent.Contains("Chrome/6"))
-            {
-                return true;
-            }
-
-            return false;
         }
+    }
+
+    private static bool DisallowsSameSiteNone(string userAgent)
+    {
+        // Cover all iOS based browsers here. This includes:
+        // - Safari on iOS 12 for iPhone, iPod Touch, iPad
+        // - WkWebview on iOS 12 for iPhone, iPod Touch, iPad
+        // - Chrome on iOS 12 for iPhone, iPod Touch, iPad
+        // All of which are broken by SameSite=None, because they use the iOS networking stack
+        if (userAgent.Contains("CPU iPhone OS 12") || userAgent.Contains("iPad; CPU OS 12"))
+        {
+            return true;
+        }
+
+        // Cover Mac OS X based browsers that use the Mac OS networking stack. This includes:
+        // - Safari on Mac OS X.
+        // This does not include:
+        // - Chrome on Mac OS X
+        // Because they do not use the Mac OS networking stack.
+        if (userAgent.Contains("Macintosh; Intel Mac OS X 10_14") &&
+            userAgent.Contains("Version/") && userAgent.Contains("Safari"))
+        {
+            return true;
+        }
+
+        // Cover Chrome 50-69, because some versions are broken by SameSite=None, 
+        // and none in this range require it.
+        // Note: this covers some pre-Chromium Edge versions, 
+        // but pre-Chromium Edge does not require SameSite=None.
+        if (userAgent.Contains("Chrome/5") || userAgent.Contains("Chrome/6"))
+        {
+            return true;
+        }
+
+        return false;
     }
 }
