@@ -3,316 +3,315 @@
 
 using StsServerIdentity.Resources;
 
-namespace Microsoft.AspNetCore.Identity
+namespace Microsoft.AspNetCore.Identity;
+
+/// <summary>
+/// Service to enable localization for application facing identity errors.
+/// </summary>
+/// <remarks>
+/// These errors are returned to controllers and are generally used as display messages to end users.
+/// </remarks>
+public class StsIdentityErrorDescriber : IdentityErrorDescriber
 {
-    /// <summary>
-    /// Service to enable localization for application facing identity errors.
-    /// </summary>
-    /// <remarks>
-    /// These errors are returned to controllers and are generally used as display messages to end users.
-    /// </remarks>
-    public class StsIdentityErrorDescriber : IdentityErrorDescriber
+    private readonly LocService _sharedLocalizer;
+
+    public StsIdentityErrorDescriber(LocService sharedLocalizer)
     {
-        private readonly LocService _sharedLocalizer;
+        _sharedLocalizer = sharedLocalizer;
+    }
+    /// <summary>
+    /// Returns the default <see cref="IdentityError"/>.
+    /// </summary>
+    /// <returns>The default <see cref="IdentityError"/>.</returns>
+    public override IdentityError DefaultError()
+    {
+        return new IdentityError
+        {
+            Code = nameof(DefaultError),
+            Description = _sharedLocalizer.GetLocalizedHtmlString("DefaultError")
+        };
+    }
 
-        public StsIdentityErrorDescriber(LocService sharedLocalizer)
+    /// <summary>
+    /// Returns an <see cref="IdentityError"/> indicating a concurrency failure.
+    /// </summary>
+    /// <returns>An <see cref="IdentityError"/> indicating a concurrency failure.</returns>
+    public override IdentityError ConcurrencyFailure()
+    {
+        return new IdentityError
         {
-            _sharedLocalizer = sharedLocalizer;
-        }
-        /// <summary>
-        /// Returns the default <see cref="IdentityError"/>.
-        /// </summary>
-        /// <returns>The default <see cref="IdentityError"/>.</returns>
-        public override IdentityError DefaultError()
-        {
-            return new IdentityError
-            {
-                Code = nameof(DefaultError),
-                Description = _sharedLocalizer.GetLocalizedHtmlString("DefaultError")
-            };
-        }
+            Code = nameof(ConcurrencyFailure),
+            Description = _sharedLocalizer.GetLocalizedHtmlString("ConcurrencyFailure")
+        };
+    }
 
-        /// <summary>
-        /// Returns an <see cref="IdentityError"/> indicating a concurrency failure.
-        /// </summary>
-        /// <returns>An <see cref="IdentityError"/> indicating a concurrency failure.</returns>
-        public override IdentityError ConcurrencyFailure()
+    /// <summary>
+    /// Returns an <see cref="IdentityError"/> indicating a password mismatch.
+    /// </summary>
+    /// <returns>An <see cref="IdentityError"/> indicating a password mismatch.</returns>
+    public override IdentityError PasswordMismatch()
+    {
+        return new IdentityError
         {
-            return new IdentityError
-            {
-                Code = nameof(ConcurrencyFailure),
-                Description = _sharedLocalizer.GetLocalizedHtmlString("ConcurrencyFailure")
-            };
-        }
+            Code = nameof(PasswordMismatch),
+            Description = _sharedLocalizer.GetLocalizedHtmlString("PasswordMismatch")
+        };
+    }
 
-        /// <summary>
-        /// Returns an <see cref="IdentityError"/> indicating a password mismatch.
-        /// </summary>
-        /// <returns>An <see cref="IdentityError"/> indicating a password mismatch.</returns>
-        public override IdentityError PasswordMismatch()
+    /// <summary>
+    /// Returns an <see cref="IdentityError"/> indicating an invalid token.
+    /// </summary>
+    /// <returns>An <see cref="IdentityError"/> indicating an invalid token.</returns>
+    public override IdentityError InvalidToken()
+    {
+        return new IdentityError
         {
-            return new IdentityError
-            {
-                Code = nameof(PasswordMismatch),
-                Description = _sharedLocalizer.GetLocalizedHtmlString("PasswordMismatch")
-            };
-        }
+            Code = nameof(InvalidToken),
+            Description = _sharedLocalizer.GetLocalizedHtmlString("InvalidToken")
+        };
+    }
 
-        /// <summary>
-        /// Returns an <see cref="IdentityError"/> indicating an invalid token.
-        /// </summary>
-        /// <returns>An <see cref="IdentityError"/> indicating an invalid token.</returns>
-        public override IdentityError InvalidToken()
+    /// <summary>
+    /// Returns an <see cref="IdentityError"/> indicating a recovery code was not redeemed.
+    /// </summary>
+    /// <returns>An <see cref="IdentityError"/> indicating a recovery code was not redeemed.</returns>
+    public override IdentityError RecoveryCodeRedemptionFailed()
+    {
+        return new IdentityError
         {
-            return new IdentityError
-            {
-                Code = nameof(InvalidToken),
-                Description = _sharedLocalizer.GetLocalizedHtmlString("InvalidToken")
-            };
-        }
+            Code = nameof(RecoveryCodeRedemptionFailed),
+            Description = _sharedLocalizer.GetLocalizedHtmlString("RecoveryCodeRedemptionFailed")
+        };
+    }
 
-        /// <summary>
-        /// Returns an <see cref="IdentityError"/> indicating a recovery code was not redeemed.
-        /// </summary>
-        /// <returns>An <see cref="IdentityError"/> indicating a recovery code was not redeemed.</returns>
-        public override IdentityError RecoveryCodeRedemptionFailed()
+    /// <summary>
+    /// Returns an <see cref="IdentityError"/> indicating an external login is already associated with an account.
+    /// </summary>
+    /// <returns>An <see cref="IdentityError"/> indicating an external login is already associated with an account.</returns>
+    public override IdentityError LoginAlreadyAssociated()
+    {
+        return new IdentityError
         {
-            return new IdentityError
-            {
-                Code = nameof(RecoveryCodeRedemptionFailed),
-                Description = _sharedLocalizer.GetLocalizedHtmlString("RecoveryCodeRedemptionFailed")
-            };
-        }
+            Code = nameof(LoginAlreadyAssociated),
+            Description = _sharedLocalizer.GetLocalizedHtmlString("LoginAlreadyAssociated")
+        };
+    }
 
-        /// <summary>
-        /// Returns an <see cref="IdentityError"/> indicating an external login is already associated with an account.
-        /// </summary>
-        /// <returns>An <see cref="IdentityError"/> indicating an external login is already associated with an account.</returns>
-        public override IdentityError LoginAlreadyAssociated()
+    /// <summary>
+    /// Returns an <see cref="IdentityError"/> indicating the specified user <paramref name="userName"/> is invalid.
+    /// </summary>
+    /// <param name="userName">The user name that is invalid.</param>
+    /// <returns>An <see cref="IdentityError"/> indicating the specified user <paramref name="userName"/> is invalid.</returns>
+    public override IdentityError InvalidUserName(string userName)
+    {
+        return new IdentityError
         {
-            return new IdentityError
-            {
-                Code = nameof(LoginAlreadyAssociated),
-                Description = _sharedLocalizer.GetLocalizedHtmlString("LoginAlreadyAssociated")
-            };
-        }
+            Code = nameof(InvalidUserName),
+            Description = _sharedLocalizer.GetLocalizedHtmlString("InvalidUserName", userName)
+        };
+    }
 
-        /// <summary>
-        /// Returns an <see cref="IdentityError"/> indicating the specified user <paramref name="userName"/> is invalid.
-        /// </summary>
-        /// <param name="userName">The user name that is invalid.</param>
-        /// <returns>An <see cref="IdentityError"/> indicating the specified user <paramref name="userName"/> is invalid.</returns>
-        public override IdentityError InvalidUserName(string userName)
+    /// <summary>
+    /// Returns an <see cref="IdentityError"/> indicating the specified <paramref name="email"/> is invalid.
+    /// </summary>
+    /// <param name="email">The email that is invalid.</param>
+    /// <returns>An <see cref="IdentityError"/> indicating the specified <paramref name="email"/> is invalid.</returns>
+    public override IdentityError InvalidEmail(string email)
+    {
+        return new IdentityError
         {
-            return new IdentityError
-            {
-                Code = nameof(InvalidUserName),
-                Description = _sharedLocalizer.GetLocalizedHtmlString("InvalidUserName", userName)
-            };
-        }
+            Code = nameof(InvalidEmail),
+            Description = _sharedLocalizer.GetLocalizedHtmlString("InvalidEmail", email)
+        };
+    }
 
-        /// <summary>
-        /// Returns an <see cref="IdentityError"/> indicating the specified <paramref name="email"/> is invalid.
-        /// </summary>
-        /// <param name="email">The email that is invalid.</param>
-        /// <returns>An <see cref="IdentityError"/> indicating the specified <paramref name="email"/> is invalid.</returns>
-        public override IdentityError InvalidEmail(string email)
+    /// <summary>
+    /// Returns an <see cref="IdentityError"/> indicating the specified <paramref name="userName"/> already exists.
+    /// </summary>
+    /// <param name="userName">The user name that already exists.</param>
+    /// <returns>An <see cref="IdentityError"/> indicating the specified <paramref name="userName"/> already exists.</returns>
+    public override IdentityError DuplicateUserName(string userName)
+    {
+        return new IdentityError
         {
-            return new IdentityError
-            {
-                Code = nameof(InvalidEmail),
-                Description = _sharedLocalizer.GetLocalizedHtmlString("InvalidEmail", email)
-            };
-        }
+            Code = nameof(DuplicateUserName),
+            Description = _sharedLocalizer.GetLocalizedHtmlString("DuplicateUserName", userName)
+        };
+    }
 
-        /// <summary>
-        /// Returns an <see cref="IdentityError"/> indicating the specified <paramref name="userName"/> already exists.
-        /// </summary>
-        /// <param name="userName">The user name that already exists.</param>
-        /// <returns>An <see cref="IdentityError"/> indicating the specified <paramref name="userName"/> already exists.</returns>
-        public override IdentityError DuplicateUserName(string userName)
+    /// <summary>
+    /// Returns an <see cref="IdentityError"/> indicating the specified <paramref name="email"/> is already associated with an account.
+    /// </summary>
+    /// <param name="email">The email that is already associated with an account.</param>
+    /// <returns>An <see cref="IdentityError"/> indicating the specified <paramref name="email"/> is already associated with an account.</returns>
+    public override IdentityError DuplicateEmail(string email)
+    {
+        return new IdentityError
         {
-            return new IdentityError
-            {
-                Code = nameof(DuplicateUserName),
-                Description = _sharedLocalizer.GetLocalizedHtmlString("DuplicateUserName", userName)
-            };
-        }
+            Code = nameof(DuplicateEmail),
+            Description = _sharedLocalizer.GetLocalizedHtmlString("DuplicateEmail", email)
+        };
+    }
 
-        /// <summary>
-        /// Returns an <see cref="IdentityError"/> indicating the specified <paramref name="email"/> is already associated with an account.
-        /// </summary>
-        /// <param name="email">The email that is already associated with an account.</param>
-        /// <returns>An <see cref="IdentityError"/> indicating the specified <paramref name="email"/> is already associated with an account.</returns>
-        public override IdentityError DuplicateEmail(string email)
+    /// <summary>
+    /// Returns an <see cref="IdentityError"/> indicating the specified <paramref name="role"/> name is invalid.
+    /// </summary>
+    /// <param name="role">The invalid role.</param>
+    /// <returns>An <see cref="IdentityError"/> indicating the specific role <paramref name="role"/> name is invalid.</returns>
+    public override IdentityError InvalidRoleName(string role)
+    {
+        return new IdentityError
         {
-            return new IdentityError
-            {
-                Code = nameof(DuplicateEmail),
-                Description = _sharedLocalizer.GetLocalizedHtmlString("DuplicateEmail", email)
-            };
-        }
+            Code = nameof(InvalidRoleName),
+            Description = _sharedLocalizer.GetLocalizedHtmlString("InvalidRoleName", role)
+        };
+    }
 
-        /// <summary>
-        /// Returns an <see cref="IdentityError"/> indicating the specified <paramref name="role"/> name is invalid.
-        /// </summary>
-        /// <param name="role">The invalid role.</param>
-        /// <returns>An <see cref="IdentityError"/> indicating the specific role <paramref name="role"/> name is invalid.</returns>
-        public override IdentityError InvalidRoleName(string role)
+    /// <summary>
+    /// Returns an <see cref="IdentityError"/> indicating the specified <paramref name="role"/> name already exists.
+    /// </summary>
+    /// <param name="role">The duplicate role.</param>
+    /// <returns>An <see cref="IdentityError"/> indicating the specific role <paramref name="role"/> name already exists.</returns>
+    public override IdentityError DuplicateRoleName(string role)
+    {
+        return new IdentityError
         {
-            return new IdentityError
-            {
-                Code = nameof(InvalidRoleName),
-                Description = _sharedLocalizer.GetLocalizedHtmlString("InvalidRoleName", role)
-            };
-        }
+            Code = nameof(DuplicateRoleName),
+            Description = _sharedLocalizer.GetLocalizedHtmlString("DuplicateRoleName", role)
+        };
+    }
 
-        /// <summary>
-        /// Returns an <see cref="IdentityError"/> indicating the specified <paramref name="role"/> name already exists.
-        /// </summary>
-        /// <param name="role">The duplicate role.</param>
-        /// <returns>An <see cref="IdentityError"/> indicating the specific role <paramref name="role"/> name already exists.</returns>
-        public override IdentityError DuplicateRoleName(string role)
+    /// <summary>
+    /// Returns an <see cref="IdentityError"/> indicating a user already has a password.
+    /// </summary>
+    /// <returns>An <see cref="IdentityError"/> indicating a user already has a password.</returns>
+    public override IdentityError UserAlreadyHasPassword()
+    {
+        return new IdentityError
         {
-            return new IdentityError
-            {
-                Code = nameof(DuplicateRoleName),
-                Description = _sharedLocalizer.GetLocalizedHtmlString("DuplicateRoleName", role)
-            };
-        }
+            Code = nameof(UserAlreadyHasPassword),
+            Description = _sharedLocalizer.GetLocalizedHtmlString("UserlreadyHasPassword")
+        };
+    }
 
-        /// <summary>
-        /// Returns an <see cref="IdentityError"/> indicating a user already has a password.
-        /// </summary>
-        /// <returns>An <see cref="IdentityError"/> indicating a user already has a password.</returns>
-        public override IdentityError UserAlreadyHasPassword()
+    /// <summary>
+    /// Returns an <see cref="IdentityError"/> indicating user lockout is not enabled.
+    /// </summary>
+    /// <returns>An <see cref="IdentityError"/> indicating user lockout is not enabled.</returns>
+    public override IdentityError UserLockoutNotEnabled()
+    {
+        return new IdentityError
         {
-            return new IdentityError
-            {
-                Code = nameof(UserAlreadyHasPassword),
-                Description = _sharedLocalizer.GetLocalizedHtmlString("UserlreadyHasPassword")
-            };
-        }
+            Code = nameof(UserLockoutNotEnabled),
+            Description = _sharedLocalizer.GetLocalizedHtmlString("UserLockoutNotEnabled")
+        };
+    }
 
-        /// <summary>
-        /// Returns an <see cref="IdentityError"/> indicating user lockout is not enabled.
-        /// </summary>
-        /// <returns>An <see cref="IdentityError"/> indicating user lockout is not enabled.</returns>
-        public override IdentityError UserLockoutNotEnabled()
+    /// <summary>
+    /// Returns an <see cref="IdentityError"/> indicating a user is already in the specified <paramref name="role"/>.
+    /// </summary>
+    /// <param name="role">The duplicate role.</param>
+    /// <returns>An <see cref="IdentityError"/> indicating a user is already in the specified <paramref name="role"/>.</returns>
+    public override IdentityError UserAlreadyInRole(string role)
+    {
+        return new IdentityError
         {
-            return new IdentityError
-            {
-                Code = nameof(UserLockoutNotEnabled),
-                Description = _sharedLocalizer.GetLocalizedHtmlString("UserLockoutNotEnabled")
-            };
-        }
+            Code = nameof(UserAlreadyInRole),
+            Description = _sharedLocalizer.GetLocalizedHtmlString("UserAlreadyInRole", role)
+        };
+    }
 
-        /// <summary>
-        /// Returns an <see cref="IdentityError"/> indicating a user is already in the specified <paramref name="role"/>.
-        /// </summary>
-        /// <param name="role">The duplicate role.</param>
-        /// <returns>An <see cref="IdentityError"/> indicating a user is already in the specified <paramref name="role"/>.</returns>
-        public override IdentityError UserAlreadyInRole(string role)
+    /// <summary>
+    /// Returns an <see cref="IdentityError"/> indicating a user is not in the specified <paramref name="role"/>.
+    /// </summary>
+    /// <param name="role">The duplicate role.</param>
+    /// <returns>An <see cref="IdentityError"/> indicating a user is not in the specified <paramref name="role"/>.</returns>
+    public override IdentityError UserNotInRole(string role)
+    {
+        return new IdentityError
         {
-            return new IdentityError
-            {
-                Code = nameof(UserAlreadyInRole),
-                Description = _sharedLocalizer.GetLocalizedHtmlString("UserAlreadyInRole", role)
-            };
-        }
+            Code = nameof(UserNotInRole),
+            Description = _sharedLocalizer.GetLocalizedHtmlString("UserNotInRole", role)
+        };
+    }
 
-        /// <summary>
-        /// Returns an <see cref="IdentityError"/> indicating a user is not in the specified <paramref name="role"/>.
-        /// </summary>
-        /// <param name="role">The duplicate role.</param>
-        /// <returns>An <see cref="IdentityError"/> indicating a user is not in the specified <paramref name="role"/>.</returns>
-        public override IdentityError UserNotInRole(string role)
+    /// <summary>
+    /// Returns an <see cref="IdentityError"/> indicating a password of the specified <paramref name="length"/> does not meet the minimum length requirements.
+    /// </summary>
+    /// <param name="length">The length that is not long enough.</param>
+    /// <returns>An <see cref="IdentityError"/> indicating a password of the specified <paramref name="length"/> does not meet the minimum length requirements.</returns>
+    public override IdentityError PasswordTooShort(int length)
+    {
+        return new IdentityError
         {
-            return new IdentityError
-            {
-                Code = nameof(UserNotInRole),
-                Description = _sharedLocalizer.GetLocalizedHtmlString("UserNotInRole", role)
-            };
-        }
+            Code = nameof(PasswordTooShort),
+            Description = _sharedLocalizer.GetLocalizedHtmlString("PasswordTooShort", length.ToString())
+        };
+    }
 
-        /// <summary>
-        /// Returns an <see cref="IdentityError"/> indicating a password of the specified <paramref name="length"/> does not meet the minimum length requirements.
-        /// </summary>
-        /// <param name="length">The length that is not long enough.</param>
-        /// <returns>An <see cref="IdentityError"/> indicating a password of the specified <paramref name="length"/> does not meet the minimum length requirements.</returns>
-        public override IdentityError PasswordTooShort(int length)
+    /// <summary>
+    /// Returns an <see cref="IdentityError"/> indicating a password does not meet the minimum number <paramref name="uniqueChars"/> of unique chars.
+    /// </summary>
+    /// <param name="uniqueChars">The number of different chars that must be used.</param>
+    /// <returns>An <see cref="IdentityError"/> indicating a password does not meet the minimum number <paramref name="uniqueChars"/> of unique chars.</returns>
+    public override IdentityError PasswordRequiresUniqueChars(int uniqueChars)
+    {
+        return new IdentityError
         {
-            return new IdentityError
-            {
-                Code = nameof(PasswordTooShort),
-                Description = _sharedLocalizer.GetLocalizedHtmlString("PasswordTooShort", length.ToString())
-            };
-        }
+            Code = nameof(PasswordRequiresUniqueChars),
+            Description = _sharedLocalizer.GetLocalizedHtmlString("PasswordRequiresUniqueChars", uniqueChars.ToString())
+        };
+    }
 
-        /// <summary>
-        /// Returns an <see cref="IdentityError"/> indicating a password does not meet the minimum number <paramref name="uniqueChars"/> of unique chars.
-        /// </summary>
-        /// <param name="uniqueChars">The number of different chars that must be used.</param>
-        /// <returns>An <see cref="IdentityError"/> indicating a password does not meet the minimum number <paramref name="uniqueChars"/> of unique chars.</returns>
-        public override IdentityError PasswordRequiresUniqueChars(int uniqueChars)
+    /// <summary>
+    /// Returns an <see cref="IdentityError"/> indicating a password entered does not contain a non-alphanumeric character, which is required by the password policy.
+    /// </summary>
+    /// <returns>An <see cref="IdentityError"/> indicating a password entered does not contain a non-alphanumeric character.</returns>
+    public override IdentityError PasswordRequiresNonAlphanumeric()
+    {
+        return new IdentityError
         {
-            return new IdentityError
-            {
-                Code = nameof(PasswordRequiresUniqueChars),
-                Description = _sharedLocalizer.GetLocalizedHtmlString("PasswordRequiresUniqueChars", uniqueChars.ToString())
-            };
-        }
+            Code = nameof(PasswordRequiresNonAlphanumeric),
+            Description = _sharedLocalizer.GetLocalizedHtmlString("PasswordRequiresNonAlphanumeric")
+        };
+    }
 
-        /// <summary>
-        /// Returns an <see cref="IdentityError"/> indicating a password entered does not contain a non-alphanumeric character, which is required by the password policy.
-        /// </summary>
-        /// <returns>An <see cref="IdentityError"/> indicating a password entered does not contain a non-alphanumeric character.</returns>
-        public override IdentityError PasswordRequiresNonAlphanumeric()
+    /// <summary>
+    /// Returns an <see cref="IdentityError"/> indicating a password entered does not contain a numeric character, which is required by the password policy.
+    /// </summary>
+    /// <returns>An <see cref="IdentityError"/> indicating a password entered does not contain a numeric character.</returns>
+    public override IdentityError PasswordRequiresDigit()
+    {
+        return new IdentityError
         {
-            return new IdentityError
-            {
-                Code = nameof(PasswordRequiresNonAlphanumeric),
-                Description = _sharedLocalizer.GetLocalizedHtmlString("PasswordRequiresNonAlphanumeric")
-            };
-        }
+            Code = nameof(PasswordRequiresDigit),
+            Description = _sharedLocalizer.GetLocalizedHtmlString("PasswordRequiresDigit")
+        };
+    }
 
-        /// <summary>
-        /// Returns an <see cref="IdentityError"/> indicating a password entered does not contain a numeric character, which is required by the password policy.
-        /// </summary>
-        /// <returns>An <see cref="IdentityError"/> indicating a password entered does not contain a numeric character.</returns>
-        public override IdentityError PasswordRequiresDigit()
+    /// <summary>
+    /// Returns an <see cref="IdentityError"/> indicating a password entered does not contain a lower case letter, which is required by the password policy.
+    /// </summary>
+    /// <returns>An <see cref="IdentityError"/> indicating a password entered does not contain a lower case letter.</returns>
+    public override IdentityError PasswordRequiresLower()
+    {
+        return new IdentityError
         {
-            return new IdentityError
-            {
-                Code = nameof(PasswordRequiresDigit),
-                Description = _sharedLocalizer.GetLocalizedHtmlString("PasswordRequiresDigit")
-            };
-        }
+            Code = nameof(PasswordRequiresLower),
+            Description = _sharedLocalizer.GetLocalizedHtmlString("PasswordRequiresLower")
+        };
+    }
 
-        /// <summary>
-        /// Returns an <see cref="IdentityError"/> indicating a password entered does not contain a lower case letter, which is required by the password policy.
-        /// </summary>
-        /// <returns>An <see cref="IdentityError"/> indicating a password entered does not contain a lower case letter.</returns>
-        public override IdentityError PasswordRequiresLower()
+    /// <summary>
+    /// Returns an <see cref="IdentityError"/> indicating a password entered does not contain an upper case letter, which is required by the password policy.
+    /// </summary>
+    /// <returns>An <see cref="IdentityError"/> indicating a password entered does not contain an upper case letter.</returns>
+    public override IdentityError PasswordRequiresUpper()
+    {
+        return new IdentityError
         {
-            return new IdentityError
-            {
-                Code = nameof(PasswordRequiresLower),
-                Description = _sharedLocalizer.GetLocalizedHtmlString("PasswordRequiresLower")
-            };
-        }
-
-        /// <summary>
-        /// Returns an <see cref="IdentityError"/> indicating a password entered does not contain an upper case letter, which is required by the password policy.
-        /// </summary>
-        /// <returns>An <see cref="IdentityError"/> indicating a password entered does not contain an upper case letter.</returns>
-        public override IdentityError PasswordRequiresUpper()
-        {
-            return new IdentityError
-            {
-                Code = nameof(PasswordRequiresUpper),
-                Description = _sharedLocalizer.GetLocalizedHtmlString("PasswordRequiresUpper")
-            };
-        }
+            Code = nameof(PasswordRequiresUpper),
+            Description = _sharedLocalizer.GetLocalizedHtmlString("PasswordRequiresUpper")
+        };
     }
 }
