@@ -1,16 +1,10 @@
-﻿using System;
+﻿namespace ApiServer.Model;
 
-namespace ApiServer.Model
+public class DataEventRecord
 {
-    public class DataEventRecord
-    {
-        public long Id { get; set; }
-        public string Name { get; set; }
-
-        public string Username { get;set;}
-
-        public string Description { get; set; }
-
-        public DateTime Timestamp { get; set; }
-    }
+    public int Id { get; set; }
+    public string Name { get; set; } = string.Empty;
+    public string Username { get; set; } = string.Empty;
+    public string Description { get; set; } = string.Empty;
+    public string Timestamp { get; set; } = string.Empty;
 }
