@@ -1,14 +1,9 @@
-﻿using System;
+﻿namespace ApiServer.Model;
 
-namespace ApiServer.Model
+public class DataEventRecordDto
 {
-    public class DataEventRecordDto
-    {
-        public long Id { get; set; }
-        public string Name { get; set; }
-
-        public string Description { get; set; }
-
-        public DateTime Timestamp { get; set; }
-    }
+    public int Id { get; set; }
+    public string Name { get; set; } = string.Empty;
+    public string Description { get; set; } = string.Empty;
+    public string Timestamp { get; set; } = string.Empty;
 }
