@@ -24,11 +24,13 @@ export class DataEventRecordsService {
     this.headers = this.headers.set('Content-Type', 'application/json');
     this.headers = this.headers.set('Accept', 'application/json');
 
-    const token = this.oidcSecurityService.getAccessToken();
-    if (token !== '') {
-      const tokenValue = 'Bearer ' + token;
-      this.headers = this.headers.append('Authorization', tokenValue);
-    }
+    this.oidcSecurityService.getAccessToken().subscribe((token) => {
+      // console.log(token)
+      if (token !== '') {
+        const tokenValue = 'Bearer ' + token;
+        this.headers = this.headers.append('Authorization', tokenValue);
+      }
+    });
   }
 
   public GetAll = (): Observable<DataEventRecord[]> => {
