@@ -15,6 +15,7 @@ export class NewsService {
   private hubConnection: HubConnection | undefined;
   private actionUrl: string;
   private headers: HttpHeaders;
+  private token: string = '';
   isAuthorized = false;
 
   constructor(
@@ -23,6 +24,7 @@ export class NewsService {
     private configuration: Configuration,
     private oidcSecurityService: OidcSecurityService
   ) {
+    console.warn('BEGIN NEWS SERVICE ...')
     this.init();
     this.actionUrl = 'https://localhost:44390/api/news/';
 
@@ -58,14 +60,6 @@ export class NewsService {
   }
 
   getAllGroups(): Observable<string[]> {
-    const token = this.oidcSecurityService.getAccessToken();
-    console.log('getAllGroups token:');
-    console.log(token);
-    if (token !== '') {
-      const tokenValue = 'Bearer ' + token;
-      this.headers = this.headers.append('Authorization', tokenValue);
-    }
-
     return this.http.get<string[]>(this.actionUrl, { headers: this.headers });
   }
 
@@ -83,44 +77,58 @@ export class NewsService {
   }
 
   private initHub() {
-    console.log('initHub');
-    const token = this.oidcSecurityService.getAccessToken();
-    let tokenValue = '';
-    if (token !== '') {
-      tokenValue = '?token=' + token;
-    }
+    console.warn('initHub');
 
-    this.hubConnection = new signalR.HubConnectionBuilder()
-      .withUrl(`${this.configuration.Server}looney${tokenValue}`)
-      .configureLogging(signalR.LogLevel.Information)
-      .build();
+    this.oidcSecurityService.isAuthenticated$.subscribe(
+      ({isAuthenticated}) => {
+        this.isAuthorized = isAuthenticated;
+        if (isAuthenticated) {
 
-    this.hubConnection.start().catch((err) => console.error(err.toString()));
+          this.oidcSecurityService.getAccessToken().subscribe((token) => {
 
-    this.hubConnection.on('Send', (newsItem: NewsItem) => {
-      this.store.dispatch(
-        newsAction.recieveNewsItemAction({ payload: newsItem })
-      );
-    });
+            let tokenValue = '';
+            this.token = token;
+            const tokenApiHeader = 'Bearer ' + this.token;
+            this.headers = this.headers.append('Authorization', tokenApiHeader);
+            console.log(tokenApiHeader)
+            tokenValue = '?token=' + token;
 
-    this.hubConnection.on('JoinGroup', (data: string) => {
-      console.log('recieved data from the hub');
-      console.log(data);
-      this.store.dispatch(
-        newsAction.recieveGroupJoinedAction({ payload: data })
-      );
-    });
+            this.hubConnection = new signalR.HubConnectionBuilder()
+              .withUrl(`${this.configuration.Server}looney${tokenValue}`)
+              .configureLogging(signalR.LogLevel.Information)
+              .build();
 
-    this.hubConnection.on('LeaveGroup', (data: string) => {
-      this.store.dispatch(newsAction.recieveGroupLeftAction({ payload: data }));
-    });
+            this.hubConnection.start().catch((err) => console.error(err.toString()));
 
-    this.hubConnection.on('History', (newsItems: NewsItem[]) => {
-      console.log('recieved history from the hub');
-      console.log(newsItems);
-      this.store.dispatch(
-        newsAction.recieveNewsGroupHistoryAction({ payload: newsItems })
-      );
-    });
+            this.hubConnection.on('Send', (newsItem: NewsItem) => {
+              this.store.dispatch(
+                newsAction.receiveNewsItemAction({ payload: newsItem })
+              );
+            });
+
+            this.hubConnection.on('JoinGroup', (data: string) => {
+              console.log('received data from the hub');
+              console.log(data);
+              this.store.dispatch(
+                newsAction.receiveGroupJoinedAction({ payload: data })
+              );
+            });
+
+            this.hubConnection.on('LeaveGroup', (data: string) => {
+              this.store.dispatch(newsAction.receiveGroupLeftAction({ payload: data }));
+            });
+
+            this.hubConnection.on('History', (newsItems: NewsItem[]) => {
+              console.log('received history from the hub');
+              console.log(newsItems);
+              this.store.dispatch(
+                newsAction.receiveNewsGroupHistoryAction({ payload: newsItems })
+              );
+            });
+          });
+
+        }
+      }
+    );
   }
 }
