@@ -1,14 +1,11 @@
-﻿using System;
+﻿namespace StsServerIdentity.Models.AccountViewModels;
 
-namespace StsServerIdentity.Models.AccountViewModels
+public class AccountOptions
 {
-    public class AccountOptions
-    {
-        public static bool AllowLocalLogin = true;
-        public static bool AllowRememberLogin = true;
-        public static TimeSpan RememberMeLoginDuration = TimeSpan.FromDays(30);
+    public static bool AllowLocalLogin = true;
+    public static bool AllowRememberLogin = true;
+    public static TimeSpan RememberMeLoginDuration = TimeSpan.FromDays(30);
 
-        public static bool ShowLogoutPrompt = true;
-        public static bool AutomaticRedirectAfterSignOut = false;
-    }
+    public static bool ShowLogoutPrompt = true;
+    public static bool AutomaticRedirectAfterSignOut = false;
 }
